% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add ggplot2 code to your tidyplot}
\usage{
add()
}
\description{
Add ggplot2 code to your tidyplot
}
\examples{
study \%>\%
  tidyplot(x = treatment, y = score, color = treatment) \%>\%
  add(ggplot2::geom_point())

}
