% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_barstack_absolute}
\alias{add_barstack_absolute}
\alias{add_barstack_relative}
\title{Add bar stack}
\usage{
add_barstack_absolute(plot, width = 0.8, reverse = FALSE, ...)

add_barstack_relative(plot, width = 0.8, reverse = FALSE, ...)
}
\arguments{
\item{plot}{A tidyplot generated with the \code{tidyplot()} function.}

\item{width}{bla}

\item{reverse}{bla}

\item{...}{bla}
}
\description{
Add bar stack
}
