% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_box}
\alias{add_box}
\title{Add boxplot}
\usage{
add_box(
  gg,
  dodge_width = NULL,
  alpha = 0.3,
  show_whiskers = TRUE,
  show_outliers = FALSE,
  box_width = 0.6,
  whiskers_width = 0.5,
  outlier.size = 0.5,
  coef = 1.5,
  outlier.shape = 19,
  linewidth = 0.25,
  ...
)
}
\arguments{
\item{gg}{bla}

\item{dodge_width}{bla}

\item{alpha}{bla}

\item{show_whiskers}{bla}

\item{show_outliers}{bla}

\item{box_width}{bla}

\item{whiskers_width}{bla}

\item{outlier.size}{bla}

\item{coef}{bla}

\item{outlier.shape}{bla}

\item{linewidth}{bla}

\item{...}{bla}
}
\description{
Add boxplot
}
