% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_boxplot}
\alias{add_boxplot}
\title{Add boxplot}
\usage{
add_boxplot(
  plot,
  dodge_width = NULL,
  saturation = 0.3,
  show_whiskers = TRUE,
  show_outliers = FALSE,
  box_width = 0.6,
  whiskers_width = 0.5,
  outlier.size = 0.5,
  coef = 1.5,
  outlier.shape = 19,
  linewidth = 0.25,
  preserve = "total",
  ...
)
}
\arguments{
\item{plot}{A tidyplot generated with the \code{tidyplot()} function.}

\item{dodge_width}{For adjusting the distance between grouped objects.}

\item{saturation}{bla}

\item{show_whiskers}{bla}

\item{show_outliers}{bla}

\item{box_width}{bla}

\item{whiskers_width}{bla}

\item{outlier.size}{bla}

\item{coef}{bla}

\item{outlier.shape}{bla}

\item{linewidth}{bla}

\item{preserve}{general}

\item{...}{bla}
}
\description{
Add boxplot
}
