% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_curve}
\alias{add_curve}
\title{Add curve}
\usage{
add_curve(
  plot,
  dodge_width = NULL,
  method = "loess",
  linewidth = 0.25,
  alpha = 0.3,
  preserve = "total",
  ...
)
}
\arguments{
\item{plot}{A tidyplot generated with the \code{tidyplot()} function.}

\item{dodge_width}{For adjusting the distance between grouped objects.}

\item{method}{bla}

\item{linewidth}{bla}

\item{alpha}{bla}

\item{preserve}{general}

\item{...}{bla}
}
\description{
Add curve
}
