% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_data_points}
\alias{add_data_points}
\alias{add_data_points_jitter}
\alias{add_data_points_beeswarm}
\title{Add data points}
\usage{
add_data_points(
  plot,
  data = all_rows(),
  shape = 19,
  size = 1,
  confetti = FALSE,
  dodge_width = NULL,
  preserve = "total",
  rasterize = FALSE,
  rasterize_dpi = 300,
  ...
)

add_data_points_jitter(
  plot,
  data = all_rows(),
  shape = 19,
  size = 1,
  confetti = FALSE,
  dodge_width = NULL,
  jitter_width = 0.2,
  jitter_height = 0,
  preserve = "total",
  rasterize = FALSE,
  rasterize_dpi = 300,
  ...
)

add_data_points_beeswarm(
  plot,
  data = all_rows(),
  shape = 19,
  size = 1,
  confetti = FALSE,
  cex = 3,
  corral = "wrap",
  corral.width = 0.5,
  dodge_width = NULL,
  preserve = "total",
  rasterize = FALSE,
  rasterize_dpi = 300,
  ...
)
}
\arguments{
\item{plot}{A tidyplot generated with the \code{tidyplot()} function.}

\item{data}{A function to subset the data to be plotted. See \code{filter_rows()} and friends.}

\item{shape}{A number representing the shape of the data points.}

\item{size}{A number representing the size of the data points.}

\item{confetti}{If \code{TRUE}, data points will get a white border. This can be
useful to deal with overplotting.}

\item{dodge_width}{For adjusting the distance between grouped objects.}

\item{preserve}{general}

\item{rasterize}{general}

\item{rasterize_dpi}{general}

\item{...}{bla}

\item{jitter_width}{For adding a small amount of random noise the to the
horizontal position of the of the data points. This can be useful to deal
with overplotting.}

\item{jitter_height}{For adding a small amount of random noise the to the
vertical position of the of the data points. This can be useful to deal
with overplotting.}

\item{cex}{Scaling for adjusting point spacing (see \code{\link[beeswarm:swarmx]{beeswarm::swarmx()}}).
Values between 1 (default) and 3 tend to work best.}

\item{corral}{\code{string}. Method used to adjust points that would be placed to
wide horizontally, default is \code{"none"}. See details below.}

\item{corral.width}{\code{numeric}. Width of the corral, default is \code{0.9}.}
}
\description{
Add data points
}
\examples{
study \%>\%
  tidyplot(x = treatment, y = score, color = treatment) \%>\%
  add_mean_bar(alpha = 0.3) \%>\%
  add_error() \%>\%
  add_data_points_beeswarm()

}
