% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_density_curve}
\alias{add_density_curve}
\title{Add density curve}
\usage{
add_density_curve(
  plot,
  bw = "nrd0",
  adjust = 1,
  kernel = "gaussian",
  n = 512,
  color = "#E37D46",
  fill = "#E37D46",
  alpha = 0.3,
  ...
)
}
\arguments{
\item{plot}{A tidyplot generated with the \code{tidyplot()} function.}

\item{bw}{bla}

\item{adjust}{bla}

\item{kernel}{bla}

\item{n}{bla}

\item{color}{bla}

\item{fill}{bla}

\item{alpha}{bla}

\item{...}{bla}
}
\description{
Add density curve
}
