% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-misc.R
\name{add_ellipse}
\alias{add_ellipse}
\title{Add ellipse}
\usage{
add_ellipse(plot, ...)
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}

\item{...}{Arguments passed on to \code{ggplot2::stat_ellipse()}.}
}
\value{
A \code{tidyplot} object.
}
\description{
Add ellipse
}
\examples{
pca |>
  tidyplot(x = pc1, y = pc2, color = group) |>
  add_data_points() |>
  add_ellipse()

pca |>
  tidyplot(x = pc1, y = pc2, color = group) |>
  add_data_points() |>
  add_ellipse(level = 0.75)

pca |>
  tidyplot(x = pc1, y = pc2, color = group) |>
  add_data_points() |>
  add_ellipse(type = "norm")

}
