% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_error_ribbon}
\alias{add_error_ribbon}
\alias{add_range_ribbon}
\alias{add_sd_ribbon}
\alias{add_ci_ribbon}
\title{Add ribbon}
\usage{
add_error_ribbon(gg, dodge_width = NULL, alpha = 0.3, color = NA, ...)

add_range_ribbon(gg, dodge_width = NULL, alpha = 0.3, color = NA, ...)

add_sd_ribbon(gg, dodge_width = NULL, alpha = 0.3, color = NA, ...)

add_ci_ribbon(gg, dodge_width = NULL, alpha = 0.3, color = NA, ...)
}
\arguments{
\item{gg}{bla}

\item{dodge_width}{bla}

\item{alpha}{bla}

\item{color}{bla}

\item{...}{bla}
}
\description{
Add ribbon
}
