% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_line}
\alias{add_line}
\alias{add_area}
\alias{add_curve}
\title{Add line}
\usage{
add_line(gg, dodge_width = NULL, linewidth = 0.25, ...)

add_area(gg, linewidth = 0.25, ...)

add_curve(
  gg,
  dodge_width = NULL,
  method = "loess",
  linewidth = 0.25,
  alpha = 0.3,
  ...
)
}
\arguments{
\item{gg}{bla}

\item{dodge_width}{bla}

\item{linewidth}{bla}

\item{...}{bla}

\item{method}{bla}

\item{alpha}{bla}
}
\description{
Add line
}
