% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_mean_bar}
\alias{add_mean_bar}
\alias{add_mean_dash}
\alias{add_mean_dot}
\alias{add_mean_value}
\alias{add_mean_line}
\alias{add_mean_area}
\title{Add mean}
\usage{
add_mean_bar(
  plot,
  dodge_width = NULL,
  width = 0.6,
  saturation = 1,
  preserve = "total",
  ...
)

add_mean_dash(
  plot,
  dodge_width = NULL,
  width = 0.6,
  linewidth = 0.25,
  preserve = "total",
  ...
)

add_mean_dot(plot, dodge_width = NULL, size = 2, preserve = "total", ...)

add_mean_value(
  plot,
  dodge_width = NULL,
  accuracy = 0.1,
  scale_cut = NULL,
  fontsize = 7,
  vjust = -0.5,
  padding_top = 0.15,
  preserve = "total",
  ...
)

add_mean_line(
  plot,
  group,
  dodge_width = NULL,
  linewidth = 0.25,
  preserve = "total",
  ...
)

add_mean_area(
  plot,
  group,
  dodge_width = NULL,
  linewidth = 0.25,
  preserve = "total",
  ...
)
}
\arguments{
\item{plot}{A tidyplot generated with the \code{tidyplot()} function.}

\item{dodge_width}{For adjusting the distance between grouped objects.}

\item{width}{bla}

\item{saturation}{bla}

\item{preserve}{general}

\item{...}{bla}

\item{linewidth}{bla}

\item{size}{bla}

\item{accuracy}{bla}

\item{scale_cut}{bla}

\item{fontsize}{bla}

\item{vjust}{bla}

\item{padding_top}{bla}

\item{group}{bla}
}
\description{
Add mean
}
