% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_pie}
\alias{add_pie}
\alias{add_donut}
\title{Add pie or donut chart}
\usage{
add_pie(plot, width = 1, reverse = FALSE, ...)

add_donut(plot, width = 1, reverse = FALSE, ...)
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}

\item{width}{Horizontal width of the plotted object (bar, error bar, boxplot,
violin plot, etc). Typical values range between \code{0} and \code{1}.}

\item{reverse}{common}

\item{...}{Arguments passed on to the \code{geom} function.}
}
\value{
A \code{tidyplot} object
}
\description{
Add pie or donut chart
}
