% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_points}
\alias{add_points}
\alias{add_jitter}
\title{Add data points}
\usage{
add_points(
  gg,
  style = c("point", "circle", "confetti"),
  data = all_rows(),
  color = NULL,
  size = 1,
  dodge_width = NULL,
  jitter_width = 0,
  jitter_height = 0,
  shape = NULL,
  preserve = "total",
  rasterize = FALSE,
  rasterize_dpi = 300,
  ...
)

add_jitter(
  gg,
  style = c("point", "circle", "confetti"),
  data = all_rows(),
  size = 1,
  dodge_width = NULL,
  jitter_width = 0.2,
  jitter_height = 0,
  shape = NULL,
  preserve = "total",
  rasterize = FALSE,
  rasterize_dpi = 300,
  ...
)
}
\arguments{
\item{gg}{bla}

\item{style}{bla}

\item{data}{bla}

\item{color}{bla}

\item{size}{bla}

\item{dodge_width}{bla}

\item{jitter_width}{bla}

\item{jitter_height}{bla}

\item{shape}{bla}

\item{preserve}{bla}

\item{rasterize}{bla}

\item{rasterize_dpi}{bla}

\item{...}{bla}
}
\description{
Add data points
}
