% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_reference_lines}
\alias{add_reference_lines}
\title{Add reference lines}
\usage{
add_reference_lines(
  gg,
  x = NULL,
  y = NULL,
  linetype = "dashed",
  linewidth = 0.25,
  ...
)
}
\arguments{
\item{gg}{bla}

\item{x}{bla}

\item{y}{bla}

\item{linetype}{bla}

\item{linewidth}{bla}

\item{...}{bla}
}
\description{
Add reference lines
}
