% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_scatter}
\alias{add_scatter}
\alias{add_jitter}
\title{Add data points}
\usage{
add_scatter(
  gg,
  subset_data = all_data(),
  dodge_width = NULL,
  point_size = 0.5,
  point_shape = 19,
  rasterize = FALSE,
  rasterize_dpi = 300,
  ...
)

add_jitter(
  gg,
  subset_data = all_data(),
  dodge_width = NULL,
  jitter_width = 0.1,
  jitter_height = 0,
  point_size = 0.5,
  point_shape = 19,
  rasterize = FALSE,
  rasterize_dpi = 300,
  ...
)
}
\arguments{
\item{gg}{bla}

\item{subset_data}{bla}

\item{dodge_width}{bla}

\item{point_size}{bla}

\item{point_shape}{bla}

\item{rasterize}{bla}

\item{rasterize_dpi}{bla}

\item{...}{bla}

\item{jitter_width}{bla}

\item{jitter_height}{bla}
}
\description{
Add data points
}
