% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_sem}
\alias{add_sem}
\alias{add_range}
\alias{add_sd}
\alias{add_ci}
\alias{add_error}
\title{Add error bar}
\usage{
add_sem(gg, dodge_width = NULL, width = 0.4, linewidth = 0.25, ...)

add_range(gg, dodge_width = NULL, width = 0.4, linewidth = 0.25, ...)

add_sd(gg, dodge_width = NULL, width = 0.4, linewidth = 0.25, ...)

add_ci(
  gg,
  conf.int = 0.95,
  dodge_width = NULL,
  width = 0.4,
  linewidth = 0.25,
  ...
)

add_error(gg, dodge_width = NULL, width = 0.4, linewidth = 0.25, ...)
}
\arguments{
\item{gg}{bla}

\item{dodge_width}{bla}

\item{width}{bla}

\item{linewidth}{bla}

\item{...}{bla}

\item{conf.int}{bla}
}
\description{
Add error bar
}
