% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_sum_bar}
\alias{add_sum_bar}
\alias{add_sum_dash}
\alias{add_sum_dot}
\alias{add_sum_value}
\alias{add_sum_line}
\alias{add_sum_area}
\title{Add sum}
\usage{
add_sum_bar(
  gg,
  dodge_width = NULL,
  width = 0.6,
  saturation = 1,
  preserve = "total",
  ...
)

add_sum_dash(
  gg,
  dodge_width = NULL,
  width = 0.6,
  linewidth = 0.25,
  preserve = "total",
  ...
)

add_sum_dot(gg, dodge_width = NULL, size = 2, preserve = "total", ...)

add_sum_value(
  gg,
  dodge_width = NULL,
  accuracy = 0.1,
  scale_cut = NULL,
  fontsize = 7,
  vjust = -0.5,
  padding_top = 0.15,
  preserve = "total",
  ...
)

add_sum_line(
  gg,
  group,
  dodge_width = NULL,
  linewidth = 0.25,
  preserve = "total",
  ...
)

add_sum_area(
  gg,
  group,
  dodge_width = NULL,
  linewidth = 0.25,
  preserve = "total",
  ...
)
}
\arguments{
\item{gg}{bla}

\item{dodge_width}{bla}

\item{width}{bla}

\item{saturation}{description}

\item{preserve}{bla}

\item{...}{description}

\item{linewidth}{bla}

\item{size}{description}

\item{accuracy}{description}

\item{scale_cut}{bla}

\item{fontsize}{description}

\item{vjust}{description}

\item{padding_top}{description}

\item{group}{bla}
}
\description{
Add sum
}
