% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-stats.R
\name{add_test_pvalue_manual}
\alias{add_test_pvalue_manual}
\alias{add_test_asterisks_manual}
\title{Add pre-computed statistics}
\usage{
add_test_pvalue_manual(
  plot,
  data,
  padding_top = 0.15,
  label = "{tidyplots::format_p_value(p, 0.0001)}",
  label.size = 7/ggplot2::.pt,
  y.position = "y.position",
  xmin = "group1",
  xmax = "group2",
  step.increase = 0.15,
  vjust = -0.25,
  bracket.nudge.y = 0.1,
  bracket.size = 0.3,
  tip.length = 0.03,
  hide.ns = FALSE,
  remove.bracket = FALSE,
  color = plot$tidyplot$ink,
  ...
)

add_test_asterisks_manual(
  plot,
  data,
  padding_top = 0.1,
  label = "p.signif",
  label.size = 10/ggplot2::.pt,
  y.position = "y.position",
  xmin = "group1",
  xmax = "group2",
  step.increase = 0.2,
  vjust = 0.3,
  bracket.nudge.y = 0.15,
  bracket.size = 0.3,
  tip.length = 0.03,
  hide.ns = TRUE,
  remove.bracket = FALSE,
  color = plot$tidyplot$ink,
  ...
)
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}

\item{data}{a data frame containing statitistical test results. The expected
default format should contain the following columns: \code{group1 | group2 |
p | y.position | etc}. \code{group1} and \code{group2} are the groups that
have been compared. \code{p} is the resulting p-value. \code{y.position} is
the y coordinates of the p-values in the plot.}

\item{padding_top}{Extra padding above the data points to accommodate the statistical comparisons.}

\item{label}{the column containing the label (e.g.: label = "p" or label =
"p.adj"), where \code{p} is the p-value. Can be also an expression that can
be formatted by the \code{\link[glue]{glue}()} package. For example, when
specifying label = "t-test, p = \{p\}", the expression \{p\} will be
replaced by its value.}

\item{label.size}{Font size of label text.}

\item{y.position}{column containing the coordinates (in data units) to be used
for absolute positioning of the label. Default value is "y.position". Can be
also a numeric vector.}

\item{xmin}{column containing the position of the left sides of the brackets.
Default value is "group1".}

\item{xmax}{(optional) column containing the position of the right sides of
the brackets. Default value is "group2". If NULL, the p-values are plotted
as a simple text.}

\item{step.increase}{numeric vector with the increase in fraction of total
height for every additional comparison to minimize overlap.}

\item{vjust}{move the text up or down relative to the bracket. Can be also a
column name available in the data.}

\item{bracket.nudge.y}{Vertical adjustment to nudge brackets by. Useful to
move up or move down the bracket. If positive value, brackets will be moved
up; if negative value, brackets are moved down.}

\item{bracket.size}{Width of the lines of the bracket.}

\item{tip.length}{numeric vector with the fraction of total height that the
bar goes down to indicate the precise column. Default is 0.03.}

\item{hide.ns}{can be logical value or a character vector.
\itemize{
\item Case when logical value. If TRUE, hide ns symbol when displaying
 significance levels. Filter is done by checking the column
 \code{p.adj.signif}, \code{p.signif}, \code{p.adj} and \code{p}.
 \item Case when character value. Possible values are "p" or "p.adj", for filtering out non significant.
 }}

\item{remove.bracket}{logical, if \code{TRUE}, brackets are removed from the
plot. Considered only in the situation, where comparisons are performed
against reference group or against "all".}

\item{color}{text and line color. Can be variable name in the data for coloring by groups.}

\item{...}{Arguments passed on to \code{ggpubr::stat_pvalue_manual()}.}
}
\value{
A \code{tidyplot} object.
}
\description{
Add pre-computed statistics
}
\details{
\itemize{
\item \code{add_test_pvalue_manual()} and \code{add_test_asterisks_manual()} use
\code{ggpubr::stat_pvalue_manual()}. Check there for additional arguments.
These functions take pre-computed statistics as input via the \code{data} parameter.
\item Known limitation: \code{add_test_pvalue_manual()} and \code{add_test_asterisks_manual()} expect a
discrete variable on the x-axis and a continuous variable on the y-axis.
To produce horizontal plots, use \code{flip_plot()}.
}
}
\examples{
# Add manual p-values from pre-computed statistics
stat_df <- data.frame(
  group1 = "A", group2 = "B",
  p = 0.025, p.signif = "*",
  y.position = 50
)
study |>
  tidyplot(x = treatment, y = score, color = treatment) |>
  add_mean_dash() |>
  add_sem_errorbar() |>
  add_data_points() |>
  add_test_pvalue_manual(data = stat_df)

# Add manual asterisks
study |>
  tidyplot(x = treatment, y = score, color = treatment) |>
  add_mean_dash() |>
  add_sem_errorbar() |>
  add_data_points() |>
  add_test_asterisks_manual(data = stat_df)

}
