% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_text_labels}
\alias{add_text_labels}
\title{Add text labels}
\usage{
add_text_labels(
  gg,
  var,
  data = all_rows(),
  fontsize = 7,
  segment.size = 0.2,
  box.padding = 0.2,
  ...
)
}
\arguments{
\item{gg}{bla}

\item{var}{bla}

\item{data}{bla}

\item{fontsize}{bla}

\item{segment.size}{bla}

\item{box.padding}{bla}

\item{...}{bla}
}
\description{
Add text labels
}
