% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_title}
\alias{add_title}
\alias{add_caption}
\title{Add plot title or caption}
\usage{
add_title(gg, title = ggplot2::waiver())

add_caption(gg, caption = ggplot2::waiver())
}
\arguments{
\item{gg}{bla}

\item{title}{bla}

\item{caption}{bla}
}
\description{
Add plot title or caption
}
