% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_violin}
\alias{add_violin}
\title{Add violin plot}
\usage{
add_violin(
  plot,
  dodge_width = NULL,
  saturation = 0.3,
  draw_quantiles = NULL,
  trim = FALSE,
  linewidth = 0.25,
  scale = "width",
  ...
)
}
\arguments{
\item{plot}{A tidyplot generated with the \code{tidyplot()} function.}

\item{dodge_width}{For adjusting the distance between grouped objects.}

\item{saturation}{bla}

\item{draw_quantiles}{bla}

\item{trim}{bla}

\item{linewidth}{bla}

\item{scale}{bla}

\item{...}{bla}
}
\description{
Add violin plot
}
