% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust_annotation}
\alias{adjust_annotation}
\title{Adjust annotation}
\usage{
adjust_annotation(
  gg,
  title = ggplot2::waiver(),
  x_axis_title = ggplot2::waiver(),
  y_axis_title = ggplot2::waiver(),
  legend_title = ggplot2::waiver(),
  caption = ggplot2::waiver(),
  ...
)
}
\arguments{
\item{gg}{bla}

\item{title}{bla}

\item{x_axis_title}{bla}

\item{y_axis_title}{bla}

\item{legend_title}{bla}

\item{caption}{bla}

\item{...}{bla}
}
\description{
Adjust annotation
}
