% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{adjust_colors}
\alias{adjust_colors}
\title{Adjust colors}
\usage{
adjust_colors(
  plot,
  new_colors = NULL,
  saturation = 1,
  labels = tidyplot_parse_labels(),
  ...
)
}
\arguments{
\item{plot}{bla}

\item{new_colors}{bla}

\item{saturation}{bla}

\item{labels}{bla}

\item{...}{bla}
}
\description{
Adjust colors
}
