% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust_data_labels}
\alias{adjust_data_labels}
\title{Adjust data labels}
\usage{
adjust_data_labels(gg, var, new_names, new_order, sort_by, reverse = FALSE)
}
\arguments{
\item{gg}{bla}

\item{var}{bla}

\item{new_names}{bla}

\item{new_order}{bla}

\item{sort_by}{bla}

\item{reverse}{bla}
}
\description{
Adjust data labels
}
