% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust_flip_plot}
\alias{adjust_flip_plot}
\title{Flip x and y axis}
\usage{
adjust_flip_plot(plot, ...)
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}

\item{...}{Arguments passed on to the \code{ggplot2::coord_flip()}.}
}
\value{
A \code{tidyplot} object
}
\description{
Flip x and y axis
}
