% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust_font}
\alias{adjust_font}
\title{Adjust font}
\usage{
adjust_font(plot, fontsize = 7, family = NULL, face = NULL, color = "black")
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}

\item{fontsize}{common}

\item{family}{Font family}

\item{face}{Font face ("plain", "italic", "bold", "bold.italic")}

\item{color}{common}
}
\value{
A \code{tidyplot} object
}
\description{
Adjust font
}
