% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust_legend}
\alias{adjust_legend}
\title{Adjust legend}
\usage{
adjust_legend(plot, title = ggplot2::waiver(), position = "right")
}
\arguments{
\item{plot}{bla}

\item{title}{bla}

\item{position}{bla}
}
\description{
Adjust legend
}
