% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust_padding}
\alias{adjust_padding}
\title{Adjust padding}
\usage{
adjust_padding(
  gg,
  top = NULL,
  right = NULL,
  bottom = NULL,
  left = NULL,
  force_y_continuous = FALSE
)
}
\arguments{
\item{gg}{bla}

\item{top}{bla}

\item{right}{bla}

\item{bottom}{bla}

\item{left}{bla}

\item{force_y_continuous}{bla}
}
\description{
Adjust padding
}
