% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust_plot_area_padding}
\alias{adjust_plot_area_padding}
\title{Adjust plot area padding}
\usage{
adjust_plot_area_padding(
  plot,
  top = NA,
  right = NA,
  bottom = NA,
  left = NA,
  force_continuous = FALSE,
  ...
)
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}

\item{top}{bla}

\item{right}{bla}

\item{bottom}{bla}

\item{left}{bla}

\item{force_continuous}{bla}

\item{...}{Arguments passed on to the \code{geom} function.}
}
\value{
A \code{tidyplot} object
}
\description{
Adjust plot area padding
}
