% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust_plot_area_size}
\alias{adjust_plot_area_size}
\title{Adjust plot area size}
\usage{
adjust_plot_area_size(plot, width = 50, height = 50, unit = "mm")
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}

\item{width}{Width of the plot area. Defaults to \code{50}.}

\item{height}{Height of the plot area. Defaults to \code{50}.}

\item{unit}{Unit of the plot area width and height. Defaults to \code{mm}.}
}
\value{
A \code{tidyplot} object
}
\description{
Adjust plot area size
}
