% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust_rotate_plot}
\alias{adjust_rotate_plot}
\title{Rotate plot by 90 degrees}
\usage{
adjust_rotate_plot(plot, ...)
}
\arguments{
\item{plot}{bla}

\item{...}{bla}
}
\description{
Rotate plot by 90 degrees
}
