% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust_variable}
\alias{adjust_variable}
\title{Adjust variable}
\usage{
adjust_variable(gg, var, new_names, new_order, sort_by, reverse = FALSE)
}
\arguments{
\item{gg}{bla}

\item{var}{bla}

\item{new_names}{bla}

\item{new_order}{bla}

\item{sort_by}{bla}

\item{reverse}{bla}
}
\description{
Adjust variable
}
