% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust_x_axis}
\alias{adjust_x_axis}
\alias{adjust_y_axis}
\title{Adjust axes}
\usage{
adjust_x_axis(
  plot,
  title = ggplot2::waiver(),
  breaks = ggplot2::waiver(),
  labels = ggplot2::waiver(),
  limits = NULL,
  padding = c(NA, NA),
  rotate_labels = FALSE,
  transformation = "identity",
  cut_short_scale = FALSE,
  force_continuous = FALSE,
  ...
)

adjust_y_axis(
  plot,
  title = ggplot2::waiver(),
  breaks = ggplot2::waiver(),
  labels = ggplot2::waiver(),
  limits = NULL,
  padding = c(NA, NA),
  rotate_labels = FALSE,
  transformation = "identity",
  cut_short_scale = FALSE,
  force_continuous = FALSE,
  ...
)
}
\arguments{
\item{plot}{bla}

\item{title}{bla}

\item{breaks}{bla}

\item{labels}{bla}

\item{limits}{bla}

\item{padding}{bla}

\item{rotate_labels}{bla}

\item{transformation}{bla}

\item{cut_short_scale}{bla}

\item{force_continuous}{bla}

\item{...}{bla}
}
\description{
Adjust axes
}
