% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{as_tidyplot}
\alias{as_tidyplot}
\title{Convert ggplot to tidyplot}
\usage{
as_tidyplot(gg, width = 50, height = 50, dodge_width = 0.8)
}
\arguments{
\item{gg}{bla}

\item{width}{bla}

\item{height}{bla}

\item{dodge_width}{bla}
}
\description{
Convert ggplot to tidyplot
}
