% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate}
\alias{climate}
\title{Climate data}
\format{
A data frame.
}
\source{
\href{https://www.noaa.gov}{National Oceanic and Atmospheric Administration}, Temperature data, weather station Hamburg Fuhlsbüttel, Germany
}
\usage{
climate
}
\description{
Climate data
}
\examples{
dplyr::glimpse(climate)
}
\keyword{datasets}
