% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{color_palettes}
\alias{color_palettes}
\title{Color palettes}
\format{
A list of 18 color palettes.
}
\source{
\code{tidyplots} package
}
\usage{
color_palettes
}
\description{
Color palettes
}
\keyword{datasets}
