% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{colors_continuous_viridis}
\alias{colors_continuous_viridis}
\alias{colors_continuous_magma}
\alias{colors_continuous_inferno}
\alias{colors_continuous_turbo}
\alias{colors_continuous_blue_pink_yellow}
\title{Continuous color schemes}
\format{
An object of class \code{tidycolor} (inherits from \code{character}) of length 265.

An object of class \code{tidycolor} (inherits from \code{character}) of length 265.

An object of class \code{tidycolor} (inherits from \code{character}) of length 265.

An object of class \code{tidycolor} (inherits from \code{character}) of length 265.

An object of class \code{tidycolor} (inherits from \code{character}) of length 11.
}
\usage{
colors_continuous_viridis

colors_continuous_magma

colors_continuous_inferno

colors_continuous_turbo

colors_continuous_blue_pink_yellow
}
\description{
Continuous color schemes
}
\details{
Previews of build-in continuous color schemes\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
  <div style='margin-bottom: 80px;'>
  <h4 style="font-family: 'Courier New', monospace;">colors_continuous_viridis</h4>
  <small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 265 colors, downsampled to 64 colors.</small>
  <p>
  <span style='background-color:#440154FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#46075AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#470D60FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#481466FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#48196BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#481E6FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#482374FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#482878FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#472E7CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#46337FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#453881FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#443C84FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#424086FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#404588FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3E4A89FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3C4F8AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3A538BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#38578CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#375B8DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#34608DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#32648EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#31688EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#2F6C8EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#2D708EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#2B748EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#2A778EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#297B8EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#277F8EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#26828EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#24878EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#228B8DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#218F8DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#20928CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#1F958BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#1F998AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#1F9E89FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#1FA187FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#21A585FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#22A983FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#25AD82FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#2BB07FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#2FB47CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#35B779FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3BBB75FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#41BE71FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4BC16CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#53C569FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#5BC864FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#63CB5FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#6CCD5AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#77D153FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#80D34DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#8AD647FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#94D741FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#9ED93BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#A8DB34FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#B5DE2BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#C0DF25FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#C9E11FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D4E21AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#DEE318FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#EBE51AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F4E61EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FDE725FF; display: inline-block; width:30px; height:20px;'></span>
  <p><p style="font-family: 'Courier New', monospace;">
  c(<br><span style='background-color:#440154FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#440154FF"</span>,<span style='background-color:#46075AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#46075AFF"</span>,<span style='background-color:#470D60FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#470D60FF"</span>,<span style='background-color:#481466FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#481466FF"</span>,<span style='background-color:#48196BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#48196BFF"</span>,<span style='background-color:#481E6FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#481E6FFF"</span>,<span style='background-color:#482374FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#482374FF"</span>,<span style='background-color:#482878FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#482878FF"</span>,<span style='background-color:#472E7CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#472E7CFF"</span>,<span style='background-color:#46337FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#46337FFF"</span>,<span style='background-color:#453881FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#453881FF"</span>,<span style='background-color:#443C84FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#443C84FF"</span>,<span style='background-color:#424086FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#424086FF"</span>,<span style='background-color:#404588FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#404588FF"</span>,<span style='background-color:#3E4A89FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3E4A89FF"</span>,<span style='background-color:#3C4F8AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3C4F8AFF"</span>,<span style='background-color:#3A538BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3A538BFF"</span>,<span style='background-color:#38578CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#38578CFF"</span>,<span style='background-color:#375B8DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#375B8DFF"</span>,<span style='background-color:#34608DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#34608DFF"</span>,<span style='background-color:#32648EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#32648EFF"</span>,<span style='background-color:#31688EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#31688EFF"</span>,<span style='background-color:#2F6C8EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#2F6C8EFF"</span>,<span style='background-color:#2D708EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#2D708EFF"</span>,<span style='background-color:#2B748EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#2B748EFF"</span>,<span style='background-color:#2A778EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#2A778EFF"</span>,<span style='background-color:#297B8EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#297B8EFF"</span>,<span style='background-color:#277F8EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#277F8EFF"</span>,<span style='background-color:#26828EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#26828EFF"</span>,<span style='background-color:#24878EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#24878EFF"</span>,<span style='background-color:#228B8DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#228B8DFF"</span>,<span style='background-color:#218F8DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#218F8DFF"</span>,<span style='background-color:#20928CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#20928CFF"</span>,<span style='background-color:#1F958BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#1F958BFF"</span>,<span style='background-color:#1F998AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#1F998AFF"</span>,<span style='background-color:#1F9E89FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#1F9E89FF"</span>,<span style='background-color:#1FA187FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#1FA187FF"</span>,<span style='background-color:#21A585FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#21A585FF"</span>,<span style='background-color:#22A983FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#22A983FF"</span>,<span style='background-color:#25AD82FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#25AD82FF"</span>,<span style='background-color:#2BB07FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#2BB07FFF"</span>,<span style='background-color:#2FB47CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#2FB47CFF"</span>,<span style='background-color:#35B779FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#35B779FF"</span>,<span style='background-color:#3BBB75FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3BBB75FF"</span>,<span style='background-color:#41BE71FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#41BE71FF"</span>,<span style='background-color:#4BC16CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#4BC16CFF"</span>,<span style='background-color:#53C569FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#53C569FF"</span>,<span style='background-color:#5BC864FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#5BC864FF"</span>,<span style='background-color:#63CB5FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#63CB5FFF"</span>,<span style='background-color:#6CCD5AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#6CCD5AFF"</span>,<span style='background-color:#77D153FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#77D153FF"</span>,<span style='background-color:#80D34DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#80D34DFF"</span>,<span style='background-color:#8AD647FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#8AD647FF"</span>,<span style='background-color:#94D741FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#94D741FF"</span>,<span style='background-color:#9ED93BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#9ED93BFF"</span>,<span style='background-color:#A8DB34FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#A8DB34FF"</span>,<span style='background-color:#B5DE2BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#B5DE2BFF"</span>,<span style='background-color:#C0DF25FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#C0DF25FF"</span>,<span style='background-color:#C9E11FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#C9E11FFF"</span>,<span style='background-color:#D4E21AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#D4E21AFF"</span>,<span style='background-color:#DEE318FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#DEE318FF"</span>,<span style='background-color:#EBE51AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#EBE51AFF"</span>,<span style='background-color:#F4E61EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F4E61EFF"</span>,<span style='background-color:#FDE725FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FDE725FF"</span>)
  </div><!--/html_preserve--><!--html_preserve-->
  <div style='margin-bottom: 80px;'>
  <h4 style="font-family: 'Courier New', monospace;">colors_continuous_magma</h4>
  <small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 265 colors, downsampled to 64 colors.</small>
  <p>
  <span style='background-color:#000004FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#020109FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#030311FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#06051BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#0A0923FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#0F0B2CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#130D34FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#180F3DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#1E1149FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#231252FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#2A115BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#301164FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#37106BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3D0F71FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#451077FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4C117AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#52137CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#58157EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#5E187FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#661A80FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#6B1D81FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#721F81FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#782281FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#7D2482FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#852781FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#8B2981FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#922B81FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#982D80FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#9E2F7FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#A6317DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#AC347CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#B3367AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#B93779FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#BF3A77FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#C53C74FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#CD4071FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D3436EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D9466BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#DF4968FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E34E64FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E95562FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#ED5A5FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F1605DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F4675CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F66D5CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F9775DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FA7E5EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FB8660FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FC8C63FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FD9467FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FE9D6CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FEA470FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FEAB75FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FEB37BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FEBA80FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FEC185FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FECA8DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FED194FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FED89AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FDDFA1FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FDE6A8FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FCEFB1FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FCF6B8FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FCFDBFFF; display: inline-block; width:30px; height:20px;'></span>
  <p><p style="font-family: 'Courier New', monospace;">
  c(<br><span style='background-color:#000004FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#000004FF"</span>,<span style='background-color:#020109FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#020109FF"</span>,<span style='background-color:#030311FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#030311FF"</span>,<span style='background-color:#06051BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#06051BFF"</span>,<span style='background-color:#0A0923FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#0A0923FF"</span>,<span style='background-color:#0F0B2CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#0F0B2CFF"</span>,<span style='background-color:#130D34FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#130D34FF"</span>,<span style='background-color:#180F3DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#180F3DFF"</span>,<span style='background-color:#1E1149FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#1E1149FF"</span>,<span style='background-color:#231252FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#231252FF"</span>,<span style='background-color:#2A115BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#2A115BFF"</span>,<span style='background-color:#301164FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#301164FF"</span>,<span style='background-color:#37106BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#37106BFF"</span>,<span style='background-color:#3D0F71FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3D0F71FF"</span>,<span style='background-color:#451077FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#451077FF"</span>,<span style='background-color:#4C117AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#4C117AFF"</span>,<span style='background-color:#52137CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#52137CFF"</span>,<span style='background-color:#58157EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#58157EFF"</span>,<span style='background-color:#5E187FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#5E187FFF"</span>,<span style='background-color:#661A80FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#661A80FF"</span>,<span style='background-color:#6B1D81FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#6B1D81FF"</span>,<span style='background-color:#721F81FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#721F81FF"</span>,<span style='background-color:#782281FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#782281FF"</span>,<span style='background-color:#7D2482FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#7D2482FF"</span>,<span style='background-color:#852781FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#852781FF"</span>,<span style='background-color:#8B2981FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#8B2981FF"</span>,<span style='background-color:#922B81FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#922B81FF"</span>,<span style='background-color:#982D80FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#982D80FF"</span>,<span style='background-color:#9E2F7FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#9E2F7FFF"</span>,<span style='background-color:#A6317DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#A6317DFF"</span>,<span style='background-color:#AC347CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#AC347CFF"</span>,<span style='background-color:#B3367AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#B3367AFF"</span>,<span style='background-color:#B93779FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#B93779FF"</span>,<span style='background-color:#BF3A77FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#BF3A77FF"</span>,<span style='background-color:#C53C74FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#C53C74FF"</span>,<span style='background-color:#CD4071FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#CD4071FF"</span>,<span style='background-color:#D3436EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#D3436EFF"</span>,<span style='background-color:#D9466BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#D9466BFF"</span>,<span style='background-color:#DF4968FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#DF4968FF"</span>,<span style='background-color:#E34E64FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#E34E64FF"</span>,<span style='background-color:#E95562FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#E95562FF"</span>,<span style='background-color:#ED5A5FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#ED5A5FFF"</span>,<span style='background-color:#F1605DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F1605DFF"</span>,<span style='background-color:#F4675CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F4675CFF"</span>,<span style='background-color:#F66D5CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F66D5CFF"</span>,<span style='background-color:#F9775DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F9775DFF"</span>,<span style='background-color:#FA7E5EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FA7E5EFF"</span>,<span style='background-color:#FB8660FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FB8660FF"</span>,<span style='background-color:#FC8C63FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FC8C63FF"</span>,<span style='background-color:#FD9467FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FD9467FF"</span>,<span style='background-color:#FE9D6CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FE9D6CFF"</span>,<span style='background-color:#FEA470FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FEA470FF"</span>,<span style='background-color:#FEAB75FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FEAB75FF"</span>,<span style='background-color:#FEB37BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FEB37BFF"</span>,<span style='background-color:#FEBA80FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FEBA80FF"</span>,<span style='background-color:#FEC185FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FEC185FF"</span>,<span style='background-color:#FECA8DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FECA8DFF"</span>,<span style='background-color:#FED194FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FED194FF"</span>,<span style='background-color:#FED89AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FED89AFF"</span>,<span style='background-color:#FDDFA1FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FDDFA1FF"</span>,<span style='background-color:#FDE6A8FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FDE6A8FF"</span>,<span style='background-color:#FCEFB1FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FCEFB1FF"</span>,<span style='background-color:#FCF6B8FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FCF6B8FF"</span>,<span style='background-color:#FCFDBFFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FCFDBFFF"</span>)
  </div><!--/html_preserve--><!--html_preserve-->
  <div style='margin-bottom: 80px;'>
  <h4 style="font-family: 'Courier New', monospace;">colors_continuous_inferno</h4>
  <small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 265 colors, downsampled to 64 colors.</small>
  <p>
  <span style='background-color:#000004FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#02010AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#040311FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#08051CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#0B0725FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#10092EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#150B37FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#1B0C41FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#230C4CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#290B54FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#300A5CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#370962FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3D0965FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#440A68FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4C0C6BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#520E6DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#59106EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#5E136EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#64156EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#6C186EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#721A6EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#781C6DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#7F1E6CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#85216BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#8D2369FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#922568FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#992766FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#9F2A63FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#A52C60FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#AD305DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#B2325AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#B83556FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#BE3853FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#C33B4FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#C83F4BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#CF4446FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D44842FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D94D3DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#DE5239FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E15635FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E65D2FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#EA632AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#ED6925FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F06F20FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F3751BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F67E14FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F7840FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F98B0BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FA9207FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FB9906FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FCA309FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FCA90EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FCB115FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FBB91EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FAC127FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F9C72FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F7D13DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F5D949FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F4E055FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F2E763FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F1EE73FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F3F587FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F6FA97FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FCFFA4FF; display: inline-block; width:30px; height:20px;'></span>
  <p><p style="font-family: 'Courier New', monospace;">
  c(<br><span style='background-color:#000004FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#000004FF"</span>,<span style='background-color:#02010AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#02010AFF"</span>,<span style='background-color:#040311FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#040311FF"</span>,<span style='background-color:#08051CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#08051CFF"</span>,<span style='background-color:#0B0725FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#0B0725FF"</span>,<span style='background-color:#10092EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#10092EFF"</span>,<span style='background-color:#150B37FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#150B37FF"</span>,<span style='background-color:#1B0C41FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#1B0C41FF"</span>,<span style='background-color:#230C4CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#230C4CFF"</span>,<span style='background-color:#290B54FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#290B54FF"</span>,<span style='background-color:#300A5CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#300A5CFF"</span>,<span style='background-color:#370962FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#370962FF"</span>,<span style='background-color:#3D0965FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3D0965FF"</span>,<span style='background-color:#440A68FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#440A68FF"</span>,<span style='background-color:#4C0C6BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#4C0C6BFF"</span>,<span style='background-color:#520E6DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#520E6DFF"</span>,<span style='background-color:#59106EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#59106EFF"</span>,<span style='background-color:#5E136EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#5E136EFF"</span>,<span style='background-color:#64156EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#64156EFF"</span>,<span style='background-color:#6C186EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#6C186EFF"</span>,<span style='background-color:#721A6EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#721A6EFF"</span>,<span style='background-color:#781C6DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#781C6DFF"</span>,<span style='background-color:#7F1E6CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#7F1E6CFF"</span>,<span style='background-color:#85216BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#85216BFF"</span>,<span style='background-color:#8D2369FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#8D2369FF"</span>,<span style='background-color:#922568FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#922568FF"</span>,<span style='background-color:#992766FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#992766FF"</span>,<span style='background-color:#9F2A63FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#9F2A63FF"</span>,<span style='background-color:#A52C60FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#A52C60FF"</span>,<span style='background-color:#AD305DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#AD305DFF"</span>,<span style='background-color:#B2325AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#B2325AFF"</span>,<span style='background-color:#B83556FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#B83556FF"</span>,<span style='background-color:#BE3853FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#BE3853FF"</span>,<span style='background-color:#C33B4FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#C33B4FFF"</span>,<span style='background-color:#C83F4BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#C83F4BFF"</span>,<span style='background-color:#CF4446FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#CF4446FF"</span>,<span style='background-color:#D44842FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#D44842FF"</span>,<span style='background-color:#D94D3DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#D94D3DFF"</span>,<span style='background-color:#DE5239FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#DE5239FF"</span>,<span style='background-color:#E15635FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#E15635FF"</span>,<span style='background-color:#E65D2FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#E65D2FFF"</span>,<span style='background-color:#EA632AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#EA632AFF"</span>,<span style='background-color:#ED6925FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#ED6925FF"</span>,<span style='background-color:#F06F20FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F06F20FF"</span>,<span style='background-color:#F3751BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F3751BFF"</span>,<span style='background-color:#F67E14FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F67E14FF"</span>,<span style='background-color:#F7840FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F7840FFF"</span>,<span style='background-color:#F98B0BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F98B0BFF"</span>,<span style='background-color:#FA9207FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FA9207FF"</span>,<span style='background-color:#FB9906FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FB9906FF"</span>,<span style='background-color:#FCA309FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FCA309FF"</span>,<span style='background-color:#FCA90EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FCA90EFF"</span>,<span style='background-color:#FCB115FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FCB115FF"</span>,<span style='background-color:#FBB91EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FBB91EFF"</span>,<span style='background-color:#FAC127FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FAC127FF"</span>,<span style='background-color:#F9C72FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F9C72FFF"</span>,<span style='background-color:#F7D13DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F7D13DFF"</span>,<span style='background-color:#F5D949FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F5D949FF"</span>,<span style='background-color:#F4E055FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F4E055FF"</span>,<span style='background-color:#F2E763FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F2E763FF"</span>,<span style='background-color:#F1EE73FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F1EE73FF"</span>,<span style='background-color:#F3F587FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F3F587FF"</span>,<span style='background-color:#F6FA97FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F6FA97FF"</span>,<span style='background-color:#FCFFA4FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FCFFA4FF"</span>)
  </div><!--/html_preserve--><!--html_preserve-->
  <div style='margin-bottom: 80px;'>
  <h4 style="font-family: 'Courier New', monospace;">colors_continuous_turbo</h4>
  <small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 265 colors, downsampled to 64 colors.</small>
  <p>
  <span style='background-color:#30123BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#351E57FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#392971FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3E378EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4141A4FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#424CB6FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4456C8FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4661D6FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#476DE6FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4777EFFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4681F7FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#458BFCFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4295FFFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3D9EFEFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#35ABF8FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#2EB4F2FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#27BEE9FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#21C6E0FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#1BCED7FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#18D8C9FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#18DEC0FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#1AE4B6FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#20EAACFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#29EFA2FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#37F492FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#44F886FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#53FA79FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#62FC6CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#71FE5FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#83FF51FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#91FF48FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#9EFD3FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#A8FC39FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#B2F936FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#BCF534FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#C8EF34FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D2E935FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#DAE336FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E2DC38FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#EAD439FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F1CA3AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F6C33AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FABA39FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FCB136FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FEA832FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FE9A2CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FE8F28FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FC8323FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F9781EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F66C19FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F15D13FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#EC540FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E74A0CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E2420AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#DC3A07FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D43305FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#CA2A04FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#C12302FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#B81D02FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#AD1701FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#A31301FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#940C01FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#880802FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#7A0403FF; display: inline-block; width:30px; height:20px;'></span>
  <p><p style="font-family: 'Courier New', monospace;">
  c(<br><span style='background-color:#30123BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#30123BFF"</span>,<span style='background-color:#351E57FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#351E57FF"</span>,<span style='background-color:#392971FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#392971FF"</span>,<span style='background-color:#3E378EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3E378EFF"</span>,<span style='background-color:#4141A4FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#4141A4FF"</span>,<span style='background-color:#424CB6FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#424CB6FF"</span>,<span style='background-color:#4456C8FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#4456C8FF"</span>,<span style='background-color:#4661D6FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#4661D6FF"</span>,<span style='background-color:#476DE6FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#476DE6FF"</span>,<span style='background-color:#4777EFFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#4777EFFF"</span>,<span style='background-color:#4681F7FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#4681F7FF"</span>,<span style='background-color:#458BFCFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#458BFCFF"</span>,<span style='background-color:#4295FFFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#4295FFFF"</span>,<span style='background-color:#3D9EFEFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3D9EFEFF"</span>,<span style='background-color:#35ABF8FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#35ABF8FF"</span>,<span style='background-color:#2EB4F2FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#2EB4F2FF"</span>,<span style='background-color:#27BEE9FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#27BEE9FF"</span>,<span style='background-color:#21C6E0FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#21C6E0FF"</span>,<span style='background-color:#1BCED7FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#1BCED7FF"</span>,<span style='background-color:#18D8C9FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#18D8C9FF"</span>,<span style='background-color:#18DEC0FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#18DEC0FF"</span>,<span style='background-color:#1AE4B6FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#1AE4B6FF"</span>,<span style='background-color:#20EAACFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#20EAACFF"</span>,<span style='background-color:#29EFA2FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#29EFA2FF"</span>,<span style='background-color:#37F492FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#37F492FF"</span>,<span style='background-color:#44F886FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#44F886FF"</span>,<span style='background-color:#53FA79FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#53FA79FF"</span>,<span style='background-color:#62FC6CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#62FC6CFF"</span>,<span style='background-color:#71FE5FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#71FE5FFF"</span>,<span style='background-color:#83FF51FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#83FF51FF"</span>,<span style='background-color:#91FF48FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#91FF48FF"</span>,<span style='background-color:#9EFD3FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#9EFD3FFF"</span>,<span style='background-color:#A8FC39FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#A8FC39FF"</span>,<span style='background-color:#B2F936FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#B2F936FF"</span>,<span style='background-color:#BCF534FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#BCF534FF"</span>,<span style='background-color:#C8EF34FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#C8EF34FF"</span>,<span style='background-color:#D2E935FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#D2E935FF"</span>,<span style='background-color:#DAE336FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#DAE336FF"</span>,<span style='background-color:#E2DC38FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#E2DC38FF"</span>,<span style='background-color:#EAD439FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#EAD439FF"</span>,<span style='background-color:#F1CA3AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F1CA3AFF"</span>,<span style='background-color:#F6C33AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F6C33AFF"</span>,<span style='background-color:#FABA39FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FABA39FF"</span>,<span style='background-color:#FCB136FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FCB136FF"</span>,<span style='background-color:#FEA832FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FEA832FF"</span>,<span style='background-color:#FE9A2CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FE9A2CFF"</span>,<span style='background-color:#FE8F28FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FE8F28FF"</span>,<span style='background-color:#FC8323FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FC8323FF"</span>,<span style='background-color:#F9781EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F9781EFF"</span>,<span style='background-color:#F66C19FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F66C19FF"</span>,<span style='background-color:#F15D13FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F15D13FF"</span>,<span style='background-color:#EC540FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#EC540FFF"</span>,<span style='background-color:#E74A0CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#E74A0CFF"</span>,<span style='background-color:#E2420AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#E2420AFF"</span>,<span style='background-color:#DC3A07FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#DC3A07FF"</span>,<span style='background-color:#D43305FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#D43305FF"</span>,<span style='background-color:#CA2A04FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#CA2A04FF"</span>,<span style='background-color:#C12302FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#C12302FF"</span>,<span style='background-color:#B81D02FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#B81D02FF"</span>,<span style='background-color:#AD1701FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#AD1701FF"</span>,<span style='background-color:#A31301FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#A31301FF"</span>,<span style='background-color:#940C01FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#940C01FF"</span>,<span style='background-color:#880802FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#880802FF"</span>,<span style='background-color:#7A0403FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#7A0403FF"</span>)
  </div><!--/html_preserve--><!--html_preserve-->
  <div style='margin-bottom: 80px;'>
  <h4 style="font-family: 'Courier New', monospace;">colors_continuous_blue_pink_yellow</h4>
  <small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 11 colors.</small>
  <p>
  <span style='background-color:#00034D; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#000F9F; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#001CEF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#241EF5; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#5823F6; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#A033E0; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E85AB1; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F1907C; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F4AF63; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FCE552; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FFFB6D; display: inline-block; width:30px; height:20px;'></span>
  <p><p style="font-family: 'Courier New', monospace;">
  c(<br><span style='background-color:#00034D; display: inline-block; padding: 2px 10px 2px 10px;'>"#00034D"</span>,<span style='background-color:#000F9F; display: inline-block; padding: 2px 10px 2px 10px;'>"#000F9F"</span>,<span style='background-color:#001CEF; display: inline-block; padding: 2px 10px 2px 10px;'>"#001CEF"</span>,<span style='background-color:#241EF5; display: inline-block; padding: 2px 10px 2px 10px;'>"#241EF5"</span>,<span style='background-color:#5823F6; display: inline-block; padding: 2px 10px 2px 10px;'>"#5823F6"</span>,<span style='background-color:#A033E0; display: inline-block; padding: 2px 10px 2px 10px;'>"#A033E0"</span>,<span style='background-color:#E85AB1; display: inline-block; padding: 2px 10px 2px 10px;'>"#E85AB1"</span>,<span style='background-color:#F1907C; display: inline-block; padding: 2px 10px 2px 10px;'>"#F1907C"</span>,<span style='background-color:#F4AF63; display: inline-block; padding: 2px 10px 2px 10px;'>"#F4AF63"</span>,<span style='background-color:#FCE552; display: inline-block; padding: 2px 10px 2px 10px;'>"#FCE552"</span>,<span style='background-color:#FFFB6D; display: inline-block; padding: 2px 10px 2px 10px;'>"#FFFB6D"</span>)
  </div><!--/html_preserve-->
}}
}
\keyword{datasets}
