% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycolor.R
\docType{data}
\name{colors_discrete_friendly}
\alias{colors_discrete_friendly}
\alias{colors_discrete_seaside}
\alias{colors_discrete_apple}
\alias{colors_discrete_friendly_long}
\alias{colors_discrete_okabeito}
\alias{colors_discrete_ibm}
\alias{colors_discrete_metro}
\alias{colors_discrete_candy}
\alias{colors_discrete_alger}
\alias{colors_discrete_rainbow}
\title{Discrete color schemes}
\format{
An object of class \code{tidycolor} (inherits from \code{character}) of length 6.

An object of class \code{tidycolor} (inherits from \code{character}) of length 5.

An object of class \code{tidycolor} (inherits from \code{character}) of length 7.

An object of class \code{tidycolor} (inherits from \code{character}) of length 7.

An object of class \code{tidycolor} (inherits from \code{character}) of length 7.

An object of class \code{tidycolor} (inherits from \code{character}) of length 5.

An object of class \code{tidycolor} (inherits from \code{character}) of length 5.

An object of class \code{tidycolor} (inherits from \code{character}) of length 5.

An object of class \code{tidycolor} (inherits from \code{character}) of length 5.

An object of class \code{tidycolor} (inherits from \code{character}) of length 9.
}
\usage{
colors_discrete_friendly

colors_discrete_seaside

colors_discrete_apple

colors_discrete_friendly_long

colors_discrete_okabeito

colors_discrete_ibm

colors_discrete_metro

colors_discrete_candy

colors_discrete_alger

colors_discrete_rainbow
}
\description{
For more information about the use of color schemes in tidyplots, check out this article:
\href{https://jbengler.github.io/tidyplots/articles/Color-schemes.html}{Color schemes}
}
\details{
The signature theme of tidyplots \code{colors_discrete_friendly} was adapted from
the \href{https://jfly.uni-koeln.de/color/}{Okabe & Ito} color palette that was designed
to work well for people with color vision deficiency.

Color schemes can be conveniently previewed by using the print method of the
\code{tidycolor} class. This will send a html preview to the RStudio Viewer pane.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_discrete_friendly
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_discrete_friendly</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 6 colors.</small><br><span style='background-color:#0072B2; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#56B4E9; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#009E73; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F5C710; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E69F00; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D55E00; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#0072B2; display: inline-block; padding: 2px 10px 2px 10px;'>"#0072B2"</span>,<span style='background-color:#56B4E9; display: inline-block; padding: 2px 10px 2px 10px;'>"#56B4E9"</span>,<span style='background-color:#009E73; display: inline-block; padding: 2px 10px 2px 10px;'>"#009E73"</span>,<span style='background-color:#F5C710; display: inline-block; padding: 2px 10px 2px 10px;'>"#F5C710"</span>,<span style='background-color:#E69F00; display: inline-block; padding: 2px 10px 2px 10px;'>"#E69F00"</span>,<span style='background-color:#D55E00; display: inline-block; padding: 2px 10px 2px 10px;'>"#D55E00"</span>)</span></div><!--/html_preserve-->
}}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_discrete_seaside
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_discrete_seaside</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 5 colors.</small><br><span style='background-color:#8ecae6; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#219ebc; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#023047; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#ffb703; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#fb8500; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#8ecae6; display: inline-block; padding: 2px 10px 2px 10px;'>"#8ecae6"</span>,<span style='background-color:#219ebc; display: inline-block; padding: 2px 10px 2px 10px;'>"#219ebc"</span>,<span style='background-color:#023047; display: inline-block; padding: 2px 10px 2px 10px;'>"#023047"</span>,<span style='background-color:#ffb703; display: inline-block; padding: 2px 10px 2px 10px;'>"#ffb703"</span>,<span style='background-color:#fb8500; display: inline-block; padding: 2px 10px 2px 10px;'>"#fb8500"</span>)</span></div><!--/html_preserve-->
}}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_discrete_apple
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_discrete_apple</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 7 colors.</small><br><span style='background-color:#ff3b30; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#ff9500; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#ffcc00; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4cd964; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#5ac8fa; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#007aff; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#5856d6; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#ff3b30; display: inline-block; padding: 2px 10px 2px 10px;'>"#ff3b30"</span>,<span style='background-color:#ff9500; display: inline-block; padding: 2px 10px 2px 10px;'>"#ff9500"</span>,<span style='background-color:#ffcc00; display: inline-block; padding: 2px 10px 2px 10px;'>"#ffcc00"</span>,<span style='background-color:#4cd964; display: inline-block; padding: 2px 10px 2px 10px;'>"#4cd964"</span>,<span style='background-color:#5ac8fa; display: inline-block; padding: 2px 10px 2px 10px;'>"#5ac8fa"</span>,<span style='background-color:#007aff; display: inline-block; padding: 2px 10px 2px 10px;'>"#007aff"</span>,<span style='background-color:#5856d6; display: inline-block; padding: 2px 10px 2px 10px;'>"#5856d6"</span>)</span></div><!--/html_preserve-->
}}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_discrete_friendly_long
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_discrete_friendly_long</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 7 colors.</small><br><span style='background-color:#CC79A7; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#0072B2; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#56B4E9; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#009E73; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F5C710; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E69F00; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D55E00; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#CC79A7; display: inline-block; padding: 2px 10px 2px 10px;'>"#CC79A7"</span>,<span style='background-color:#0072B2; display: inline-block; padding: 2px 10px 2px 10px;'>"#0072B2"</span>,<span style='background-color:#56B4E9; display: inline-block; padding: 2px 10px 2px 10px;'>"#56B4E9"</span>,<span style='background-color:#009E73; display: inline-block; padding: 2px 10px 2px 10px;'>"#009E73"</span>,<span style='background-color:#F5C710; display: inline-block; padding: 2px 10px 2px 10px;'>"#F5C710"</span>,<span style='background-color:#E69F00; display: inline-block; padding: 2px 10px 2px 10px;'>"#E69F00"</span>,<span style='background-color:#D55E00; display: inline-block; padding: 2px 10px 2px 10px;'>"#D55E00"</span>)</span></div><!--/html_preserve-->
}}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_discrete_okabeito
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_discrete_okabeito</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 7 colors.</small><br><span style='background-color:#E69F00; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#56B4E9; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#009E73; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F0E442; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#0072B2; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D55E00; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#CC79A7; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#E69F00; display: inline-block; padding: 2px 10px 2px 10px;'>"#E69F00"</span>,<span style='background-color:#56B4E9; display: inline-block; padding: 2px 10px 2px 10px;'>"#56B4E9"</span>,<span style='background-color:#009E73; display: inline-block; padding: 2px 10px 2px 10px;'>"#009E73"</span>,<span style='background-color:#F0E442; display: inline-block; padding: 2px 10px 2px 10px;'>"#F0E442"</span>,<span style='background-color:#0072B2; display: inline-block; padding: 2px 10px 2px 10px;'>"#0072B2"</span>,<span style='background-color:#D55E00; display: inline-block; padding: 2px 10px 2px 10px;'>"#D55E00"</span>,<span style='background-color:#CC79A7; display: inline-block; padding: 2px 10px 2px 10px;'>"#CC79A7"</span>)</span></div><!--/html_preserve-->
}}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_discrete_ibm
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_discrete_ibm</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 5 colors.</small><br><span style='background-color:#5B8DFE; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#725DEE; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#DD227D; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FE5F00; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FFB109; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#5B8DFE; display: inline-block; padding: 2px 10px 2px 10px;'>"#5B8DFE"</span>,<span style='background-color:#725DEE; display: inline-block; padding: 2px 10px 2px 10px;'>"#725DEE"</span>,<span style='background-color:#DD227D; display: inline-block; padding: 2px 10px 2px 10px;'>"#DD227D"</span>,<span style='background-color:#FE5F00; display: inline-block; padding: 2px 10px 2px 10px;'>"#FE5F00"</span>,<span style='background-color:#FFB109; display: inline-block; padding: 2px 10px 2px 10px;'>"#FFB109"</span>)</span></div><!--/html_preserve-->
}}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_discrete_metro
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_discrete_metro</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 5 colors.</small><br><span style='background-color:#4DACD6; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4FAE62; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F6C54D; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E37D46; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#C02D45; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#4DACD6; display: inline-block; padding: 2px 10px 2px 10px;'>"#4DACD6"</span>,<span style='background-color:#4FAE62; display: inline-block; padding: 2px 10px 2px 10px;'>"#4FAE62"</span>,<span style='background-color:#F6C54D; display: inline-block; padding: 2px 10px 2px 10px;'>"#F6C54D"</span>,<span style='background-color:#E37D46; display: inline-block; padding: 2px 10px 2px 10px;'>"#E37D46"</span>,<span style='background-color:#C02D45; display: inline-block; padding: 2px 10px 2px 10px;'>"#C02D45"</span>)</span></div><!--/html_preserve-->
}}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_discrete_candy
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_discrete_candy</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 5 colors.</small><br><span style='background-color:#9b5de5; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#f15bb5; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#fee440; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#00bbf9; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#00f5d4; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#9b5de5; display: inline-block; padding: 2px 10px 2px 10px;'>"#9b5de5"</span>,<span style='background-color:#f15bb5; display: inline-block; padding: 2px 10px 2px 10px;'>"#f15bb5"</span>,<span style='background-color:#fee440; display: inline-block; padding: 2px 10px 2px 10px;'>"#fee440"</span>,<span style='background-color:#00bbf9; display: inline-block; padding: 2px 10px 2px 10px;'>"#00bbf9"</span>,<span style='background-color:#00f5d4; display: inline-block; padding: 2px 10px 2px 10px;'>"#00f5d4"</span>)</span></div><!--/html_preserve-->
}}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_discrete_alger
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_discrete_alger</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 5 colors.</small><br><span style='background-color:#000000; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#1A5B5B; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#ACC8BE; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F4AB5C; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D1422F; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#000000; display: inline-block; padding: 2px 10px 2px 10px;'>"#000000"</span>,<span style='background-color:#1A5B5B; display: inline-block; padding: 2px 10px 2px 10px;'>"#1A5B5B"</span>,<span style='background-color:#ACC8BE; display: inline-block; padding: 2px 10px 2px 10px;'>"#ACC8BE"</span>,<span style='background-color:#F4AB5C; display: inline-block; padding: 2px 10px 2px 10px;'>"#F4AB5C"</span>,<span style='background-color:#D1422F; display: inline-block; padding: 2px 10px 2px 10px;'>"#D1422F"</span>)</span></div><!--/html_preserve-->
}}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_discrete_rainbow
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_discrete_rainbow</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 9 colors.</small><br><span style='background-color:#FF7777; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FFAB74; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FFE577; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#DBF47B; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#91E480; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#7CC9E5; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#7DA8E6; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#887DE6; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#BC7BE4; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#FF7777; display: inline-block; padding: 2px 10px 2px 10px;'>"#FF7777"</span>,<span style='background-color:#FFAB74; display: inline-block; padding: 2px 10px 2px 10px;'>"#FFAB74"</span>,<span style='background-color:#FFE577; display: inline-block; padding: 2px 10px 2px 10px;'>"#FFE577"</span>,<span style='background-color:#DBF47B; display: inline-block; padding: 2px 10px 2px 10px;'>"#DBF47B"</span>,<span style='background-color:#91E480; display: inline-block; padding: 2px 10px 2px 10px;'>"#91E480"</span>,<span style='background-color:#7CC9E5; display: inline-block; padding: 2px 10px 2px 10px;'>"#7CC9E5"</span>,<span style='background-color:#7DA8E6; display: inline-block; padding: 2px 10px 2px 10px;'>"#7DA8E6"</span>,<span style='background-color:#887DE6; display: inline-block; padding: 2px 10px 2px 10px;'>"#887DE6"</span>,<span style='background-color:#BC7BE4; display: inline-block; padding: 2px 10px 2px 10px;'>"#BC7BE4"</span>)</span></div><!--/html_preserve-->
}}
}
\keyword{datasets}
