% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{colors_discrete_metro}
\alias{colors_discrete_metro}
\alias{colors_discrete_circle}
\alias{colors_discrete_candy}
\alias{colors_discrete_pastel}
\alias{colors_discrete_seaside}
\title{Discrete color schemes}
\format{
An object of class \code{tidycolor} (inherits from \code{character}) of length 5.

An object of class \code{tidycolor} (inherits from \code{character}) of length 6.

An object of class \code{tidycolor} (inherits from \code{character}) of length 5.

An object of class \code{tidycolor} (inherits from \code{character}) of length 5.

An object of class \code{tidycolor} (inherits from \code{character}) of length 5.
}
\usage{
colors_discrete_metro

colors_discrete_circle

colors_discrete_candy

colors_discrete_pastel

colors_discrete_seaside
}
\description{
Discrete color schemes
}
\details{
Previews of build-in discrete color schemes\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
  <div style='margin-bottom: 80px;'>
  <h4 style="font-family: 'Courier New', monospace;">colors_discrete_metro</h4>
  <small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 5 colors.</small>
  <p>
  <span style='background-color:#4DACD6; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4FAE62; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F6C54D; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E37D46; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#C02D45; display: inline-block; width:30px; height:20px;'></span>
  <p><p style="font-family: 'Courier New', monospace;">
  c(<br><span style='background-color:#4DACD6; display: inline-block; padding: 2px 10px 2px 10px;'>"#4DACD6"</span>,<span style='background-color:#4FAE62; display: inline-block; padding: 2px 10px 2px 10px;'>"#4FAE62"</span>,<span style='background-color:#F6C54D; display: inline-block; padding: 2px 10px 2px 10px;'>"#F6C54D"</span>,<span style='background-color:#E37D46; display: inline-block; padding: 2px 10px 2px 10px;'>"#E37D46"</span>,<span style='background-color:#C02D45; display: inline-block; padding: 2px 10px 2px 10px;'>"#C02D45"</span>)
  </div><!--/html_preserve--><!--html_preserve-->
  <div style='margin-bottom: 80px;'>
  <h4 style="font-family: 'Courier New', monospace;">colors_discrete_circle</h4>
  <small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 6 colors.</small>
  <p>
  <span style='background-color:#939393; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#644296; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F08533; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D1352C; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#559E3F; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3B78B0; display: inline-block; width:30px; height:20px;'></span>
  <p><p style="font-family: 'Courier New', monospace;">
  c(<br><span style='background-color:#939393; display: inline-block; padding: 2px 10px 2px 10px;'>"#939393"</span>,<span style='background-color:#644296; display: inline-block; padding: 2px 10px 2px 10px;'>"#644296"</span>,<span style='background-color:#F08533; display: inline-block; padding: 2px 10px 2px 10px;'>"#F08533"</span>,<span style='background-color:#D1352C; display: inline-block; padding: 2px 10px 2px 10px;'>"#D1352C"</span>,<span style='background-color:#559E3F; display: inline-block; padding: 2px 10px 2px 10px;'>"#559E3F"</span>,<span style='background-color:#3B78B0; display: inline-block; padding: 2px 10px 2px 10px;'>"#3B78B0"</span>)
  </div><!--/html_preserve--><!--html_preserve-->
  <div style='margin-bottom: 80px;'>
  <h4 style="font-family: 'Courier New', monospace;">colors_discrete_candy</h4>
  <small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 5 colors.</small>
  <p>
  <span style='background-color:#9b5de5; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#f15bb5; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#fee440; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#00bbf9; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#00f5d4; display: inline-block; width:30px; height:20px;'></span>
  <p><p style="font-family: 'Courier New', monospace;">
  c(<br><span style='background-color:#9b5de5; display: inline-block; padding: 2px 10px 2px 10px;'>"#9b5de5"</span>,<span style='background-color:#f15bb5; display: inline-block; padding: 2px 10px 2px 10px;'>"#f15bb5"</span>,<span style='background-color:#fee440; display: inline-block; padding: 2px 10px 2px 10px;'>"#fee440"</span>,<span style='background-color:#00bbf9; display: inline-block; padding: 2px 10px 2px 10px;'>"#00bbf9"</span>,<span style='background-color:#00f5d4; display: inline-block; padding: 2px 10px 2px 10px;'>"#00f5d4"</span>)
  </div><!--/html_preserve--><!--html_preserve-->
  <div style='margin-bottom: 80px;'>
  <h4 style="font-family: 'Courier New', monospace;">colors_discrete_pastel</h4>
  <small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 5 colors.</small>
  <p>
  <span style='background-color:#cdb4db; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#ffc8dd; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#ffafcc; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#bde0fe; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#a2d2ff; display: inline-block; width:30px; height:20px;'></span>
  <p><p style="font-family: 'Courier New', monospace;">
  c(<br><span style='background-color:#cdb4db; display: inline-block; padding: 2px 10px 2px 10px;'>"#cdb4db"</span>,<span style='background-color:#ffc8dd; display: inline-block; padding: 2px 10px 2px 10px;'>"#ffc8dd"</span>,<span style='background-color:#ffafcc; display: inline-block; padding: 2px 10px 2px 10px;'>"#ffafcc"</span>,<span style='background-color:#bde0fe; display: inline-block; padding: 2px 10px 2px 10px;'>"#bde0fe"</span>,<span style='background-color:#a2d2ff; display: inline-block; padding: 2px 10px 2px 10px;'>"#a2d2ff"</span>)
  </div><!--/html_preserve--><!--html_preserve-->
  <div style='margin-bottom: 80px;'>
  <h4 style="font-family: 'Courier New', monospace;">colors_discrete_seaside</h4>
  <small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 5 colors.</small>
  <p>
  <span style='background-color:#8ecae6; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#219ebc; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#023047; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#ffb703; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#fb8500; display: inline-block; width:30px; height:20px;'></span>
  <p><p style="font-family: 'Courier New', monospace;">
  c(<br><span style='background-color:#8ecae6; display: inline-block; padding: 2px 10px 2px 10px;'>"#8ecae6"</span>,<span style='background-color:#219ebc; display: inline-block; padding: 2px 10px 2px 10px;'>"#219ebc"</span>,<span style='background-color:#023047; display: inline-block; padding: 2px 10px 2px 10px;'>"#023047"</span>,<span style='background-color:#ffb703; display: inline-block; padding: 2px 10px 2px 10px;'>"#ffb703"</span>,<span style='background-color:#fb8500; display: inline-block; padding: 2px 10px 2px 10px;'>"#fb8500"</span>)
  </div><!--/html_preserve-->
}}
}
\keyword{datasets}
