% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycolor.R
\docType{data}
\name{colors_diverging_blue2brown}
\alias{colors_diverging_blue2brown}
\alias{colors_diverging_blue2red}
\alias{colors_diverging_BuRd}
\alias{colors_diverging_BuYlRd}
\title{Diverging color schemes}
\format{
An object of class \code{tidycolor} (inherits from \code{character}) of length 17.

An object of class \code{tidycolor} (inherits from \code{character}) of length 9.

An object of class \code{tidycolor} (inherits from \code{character}) of length 9.

An object of class \code{tidycolor} (inherits from \code{character}) of length 9.
}
\usage{
colors_diverging_blue2brown

colors_diverging_blue2red

colors_diverging_BuRd

colors_diverging_BuYlRd
}
\description{
Diverging color schemes
}
\details{
Color schemes can be conveniently previewed by using the print method of the
\code{tidycolor} class. This will send a html preview to the RStudio Viewer pane
that looks like this:\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
  <div style='margin-bottom: 80px;'>
  <h4 style="font-family: 'Courier New', monospace;">colors_diverging_blue2brown</h4>
  <small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 17 colors.</small>
  <p>
  <span style='background-color:#1961A5; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#2671B5; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#2D80BF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#268CC9; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#119DD8; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#00B2EB; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#66C5EF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#C4E5F8; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FEFCF6; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FDEEB8; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FCDD67; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F6C445; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E78B43; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#DD5642; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#DB3E34; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#CA3632; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#B3322E; display: inline-block; width:30px; height:20px;'></span>
  <p><p style="font-family: 'Courier New', monospace;">
  c(<br><span style='background-color:#1961A5; display: inline-block; padding: 2px 10px 2px 10px;'>"#1961A5"</span>,<span style='background-color:#2671B5; display: inline-block; padding: 2px 10px 2px 10px;'>"#2671B5"</span>,<span style='background-color:#2D80BF; display: inline-block; padding: 2px 10px 2px 10px;'>"#2D80BF"</span>,<span style='background-color:#268CC9; display: inline-block; padding: 2px 10px 2px 10px;'>"#268CC9"</span>,<span style='background-color:#119DD8; display: inline-block; padding: 2px 10px 2px 10px;'>"#119DD8"</span>,<span style='background-color:#00B2EB; display: inline-block; padding: 2px 10px 2px 10px;'>"#00B2EB"</span>,<span style='background-color:#66C5EF; display: inline-block; padding: 2px 10px 2px 10px;'>"#66C5EF"</span>,<span style='background-color:#C4E5F8; display: inline-block; padding: 2px 10px 2px 10px;'>"#C4E5F8"</span>,<span style='background-color:#FEFCF6; display: inline-block; padding: 2px 10px 2px 10px;'>"#FEFCF6"</span>,<span style='background-color:#FDEEB8; display: inline-block; padding: 2px 10px 2px 10px;'>"#FDEEB8"</span>,<span style='background-color:#FCDD67; display: inline-block; padding: 2px 10px 2px 10px;'>"#FCDD67"</span>,<span style='background-color:#F6C445; display: inline-block; padding: 2px 10px 2px 10px;'>"#F6C445"</span>,<span style='background-color:#E78B43; display: inline-block; padding: 2px 10px 2px 10px;'>"#E78B43"</span>,<span style='background-color:#DD5642; display: inline-block; padding: 2px 10px 2px 10px;'>"#DD5642"</span>,<span style='background-color:#DB3E34; display: inline-block; padding: 2px 10px 2px 10px;'>"#DB3E34"</span>,<span style='background-color:#CA3632; display: inline-block; padding: 2px 10px 2px 10px;'>"#CA3632"</span>,<span style='background-color:#B3322E; display: inline-block; padding: 2px 10px 2px 10px;'>"#B3322E"</span>)
  </div><!--/html_preserve--><!--html_preserve-->
  <div style='margin-bottom: 80px;'>
  <h4 style="font-family: 'Courier New', monospace;">colors_diverging_blue2red</h4>
  <small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 9 colors.</small>
  <p>
  <span style='background-color:#0000FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3F3FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#7F7FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#BFBFFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FFFFFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FFBFBF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FF7F7F; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FF3F3F; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FF0000; display: inline-block; width:30px; height:20px;'></span>
  <p><p style="font-family: 'Courier New', monospace;">
  c(<br><span style='background-color:#0000FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#0000FF"</span>,<span style='background-color:#3F3FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3F3FFF"</span>,<span style='background-color:#7F7FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#7F7FFF"</span>,<span style='background-color:#BFBFFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#BFBFFF"</span>,<span style='background-color:#FFFFFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FFFFFF"</span>,<span style='background-color:#FFBFBF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FFBFBF"</span>,<span style='background-color:#FF7F7F; display: inline-block; padding: 2px 10px 2px 10px;'>"#FF7F7F"</span>,<span style='background-color:#FF3F3F; display: inline-block; padding: 2px 10px 2px 10px;'>"#FF3F3F"</span>,<span style='background-color:#FF0000; display: inline-block; padding: 2px 10px 2px 10px;'>"#FF0000"</span>)
  </div><!--/html_preserve--><!--html_preserve-->
  <div style='margin-bottom: 80px;'>
  <h4 style="font-family: 'Courier New', monospace;">colors_diverging_BuRd</h4>
  <small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 9 colors.</small>
  <p>
  <span style='background-color:#2166AC; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4393C3; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#92C5DE; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D1E5F0; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F7F7F7; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FDDBC7; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F4A582; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D6604D; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#B2182B; display: inline-block; width:30px; height:20px;'></span>
  <p><p style="font-family: 'Courier New', monospace;">
  c(<br><span style='background-color:#2166AC; display: inline-block; padding: 2px 10px 2px 10px;'>"#2166AC"</span>,<span style='background-color:#4393C3; display: inline-block; padding: 2px 10px 2px 10px;'>"#4393C3"</span>,<span style='background-color:#92C5DE; display: inline-block; padding: 2px 10px 2px 10px;'>"#92C5DE"</span>,<span style='background-color:#D1E5F0; display: inline-block; padding: 2px 10px 2px 10px;'>"#D1E5F0"</span>,<span style='background-color:#F7F7F7; display: inline-block; padding: 2px 10px 2px 10px;'>"#F7F7F7"</span>,<span style='background-color:#FDDBC7; display: inline-block; padding: 2px 10px 2px 10px;'>"#FDDBC7"</span>,<span style='background-color:#F4A582; display: inline-block; padding: 2px 10px 2px 10px;'>"#F4A582"</span>,<span style='background-color:#D6604D; display: inline-block; padding: 2px 10px 2px 10px;'>"#D6604D"</span>,<span style='background-color:#B2182B; display: inline-block; padding: 2px 10px 2px 10px;'>"#B2182B"</span>)
  </div><!--/html_preserve--><!--html_preserve-->
  <div style='margin-bottom: 80px;'>
  <h4 style="font-family: 'Courier New', monospace;">colors_diverging_BuYlRd</h4>
  <small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 9 colors.</small>
  <p>
  <span style='background-color:#4575B4; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#74ADD1; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#ABD9E9; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E0F3F8; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FFFFBF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FEE090; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FDAE61; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F46D43; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D73027; display: inline-block; width:30px; height:20px;'></span>
  <p><p style="font-family: 'Courier New', monospace;">
  c(<br><span style='background-color:#4575B4; display: inline-block; padding: 2px 10px 2px 10px;'>"#4575B4"</span>,<span style='background-color:#74ADD1; display: inline-block; padding: 2px 10px 2px 10px;'>"#74ADD1"</span>,<span style='background-color:#ABD9E9; display: inline-block; padding: 2px 10px 2px 10px;'>"#ABD9E9"</span>,<span style='background-color:#E0F3F8; display: inline-block; padding: 2px 10px 2px 10px;'>"#E0F3F8"</span>,<span style='background-color:#FFFFBF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FFFFBF"</span>,<span style='background-color:#FEE090; display: inline-block; padding: 2px 10px 2px 10px;'>"#FEE090"</span>,<span style='background-color:#FDAE61; display: inline-block; padding: 2px 10px 2px 10px;'>"#FDAE61"</span>,<span style='background-color:#F46D43; display: inline-block; padding: 2px 10px 2px 10px;'>"#F46D43"</span>,<span style='background-color:#D73027; display: inline-block; padding: 2px 10px 2px 10px;'>"#D73027"</span>)
  </div><!--/html_preserve-->
}}
}
\keyword{datasets}
