% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycolor.R
\docType{data}
\name{colors_diverging_blue2red}
\alias{colors_diverging_blue2red}
\alias{colors_diverging_blue2brown}
\alias{colors_diverging_BuRd}
\alias{colors_diverging_BuYlRd}
\alias{colors_diverging_spectral}
\alias{colors_diverging_icefire}
\title{Diverging color schemes}
\format{
An object of class \code{tidycolor} (inherits from \code{character}) of length 17.

An object of class \code{tidycolor} (inherits from \code{character}) of length 17.

An object of class \code{tidycolor} (inherits from \code{character}) of length 11.

An object of class \code{tidycolor} (inherits from \code{character}) of length 11.

An object of class \code{tidycolor} (inherits from \code{character}) of length 96.

An object of class \code{tidycolor} (inherits from \code{character}) of length 96.
}
\usage{
colors_diverging_blue2red

colors_diverging_blue2brown

colors_diverging_BuRd

colors_diverging_BuYlRd

colors_diverging_spectral

colors_diverging_icefire
}
\description{
For more information about the use of color schemes in tidyplots, check out this article:
\href{https://jbengler.github.io/tidyplots/articles/Color-schemes.html}{Color schemes}
}
\details{
Color schemes can be conveniently previewed by using the print method of the
\code{tidycolor} class. This will send a html preview to the RStudio Viewer pane.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_diverging_blue2red
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_diverging_blue2red</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 17 colors.</small><br><span style='background-color:#0000FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#1F1FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3F3FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#5F5FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#7F7FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#9F9FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#BFBFFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#DFDFFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FFFFFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FFDFDF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FFBFBF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FF9F9F; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FF7F7F; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FF5F5F; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FF3F3F; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FF1F1F; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FF0000; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#0000FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#0000FF"</span>,<span style='background-color:#1F1FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#1F1FFF"</span>,<span style='background-color:#3F3FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3F3FFF"</span>,<span style='background-color:#5F5FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#5F5FFF"</span>,<span style='background-color:#7F7FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#7F7FFF"</span>,<span style='background-color:#9F9FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#9F9FFF"</span>,<span style='background-color:#BFBFFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#BFBFFF"</span>,<span style='background-color:#DFDFFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#DFDFFF"</span>,<span style='background-color:#FFFFFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FFFFFF"</span>,<span style='background-color:#FFDFDF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FFDFDF"</span>,<span style='background-color:#FFBFBF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FFBFBF"</span>,<span style='background-color:#FF9F9F; display: inline-block; padding: 2px 10px 2px 10px;'>"#FF9F9F"</span>,<span style='background-color:#FF7F7F; display: inline-block; padding: 2px 10px 2px 10px;'>"#FF7F7F"</span>,<span style='background-color:#FF5F5F; display: inline-block; padding: 2px 10px 2px 10px;'>"#FF5F5F"</span>,<span style='background-color:#FF3F3F; display: inline-block; padding: 2px 10px 2px 10px;'>"#FF3F3F"</span>,<span style='background-color:#FF1F1F; display: inline-block; padding: 2px 10px 2px 10px;'>"#FF1F1F"</span>,<span style='background-color:#FF0000; display: inline-block; padding: 2px 10px 2px 10px;'>"#FF0000"</span>)</span></div><!--/html_preserve-->
}}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_diverging_blue2brown
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_diverging_blue2brown</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 17 colors.</small><br><span style='background-color:#1961A5; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#2671B5; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#2D80BF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#268CC9; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#119DD8; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#00B2EB; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#66C5EF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#C4E5F8; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FEFCF6; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FDEEB8; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FCDD67; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F6C445; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E78B43; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#DD5642; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#DB3E34; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#CA3632; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#B3322E; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#1961A5; display: inline-block; padding: 2px 10px 2px 10px;'>"#1961A5"</span>,<span style='background-color:#2671B5; display: inline-block; padding: 2px 10px 2px 10px;'>"#2671B5"</span>,<span style='background-color:#2D80BF; display: inline-block; padding: 2px 10px 2px 10px;'>"#2D80BF"</span>,<span style='background-color:#268CC9; display: inline-block; padding: 2px 10px 2px 10px;'>"#268CC9"</span>,<span style='background-color:#119DD8; display: inline-block; padding: 2px 10px 2px 10px;'>"#119DD8"</span>,<span style='background-color:#00B2EB; display: inline-block; padding: 2px 10px 2px 10px;'>"#00B2EB"</span>,<span style='background-color:#66C5EF; display: inline-block; padding: 2px 10px 2px 10px;'>"#66C5EF"</span>,<span style='background-color:#C4E5F8; display: inline-block; padding: 2px 10px 2px 10px;'>"#C4E5F8"</span>,<span style='background-color:#FEFCF6; display: inline-block; padding: 2px 10px 2px 10px;'>"#FEFCF6"</span>,<span style='background-color:#FDEEB8; display: inline-block; padding: 2px 10px 2px 10px;'>"#FDEEB8"</span>,<span style='background-color:#FCDD67; display: inline-block; padding: 2px 10px 2px 10px;'>"#FCDD67"</span>,<span style='background-color:#F6C445; display: inline-block; padding: 2px 10px 2px 10px;'>"#F6C445"</span>,<span style='background-color:#E78B43; display: inline-block; padding: 2px 10px 2px 10px;'>"#E78B43"</span>,<span style='background-color:#DD5642; display: inline-block; padding: 2px 10px 2px 10px;'>"#DD5642"</span>,<span style='background-color:#DB3E34; display: inline-block; padding: 2px 10px 2px 10px;'>"#DB3E34"</span>,<span style='background-color:#CA3632; display: inline-block; padding: 2px 10px 2px 10px;'>"#CA3632"</span>,<span style='background-color:#B3322E; display: inline-block; padding: 2px 10px 2px 10px;'>"#B3322E"</span>)</span></div><!--/html_preserve-->
}}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_diverging_BuRd
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_diverging_BuRd</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 11 colors.</small><br><span style='background-color:#053061; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#2166AC; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4393C3; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#92C5DE; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D1E5F0; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F7F7F7; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FDDBC7; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F4A582; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D6604D; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#B2182B; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#67001F; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#053061; display: inline-block; padding: 2px 10px 2px 10px;'>"#053061"</span>,<span style='background-color:#2166AC; display: inline-block; padding: 2px 10px 2px 10px;'>"#2166AC"</span>,<span style='background-color:#4393C3; display: inline-block; padding: 2px 10px 2px 10px;'>"#4393C3"</span>,<span style='background-color:#92C5DE; display: inline-block; padding: 2px 10px 2px 10px;'>"#92C5DE"</span>,<span style='background-color:#D1E5F0; display: inline-block; padding: 2px 10px 2px 10px;'>"#D1E5F0"</span>,<span style='background-color:#F7F7F7; display: inline-block; padding: 2px 10px 2px 10px;'>"#F7F7F7"</span>,<span style='background-color:#FDDBC7; display: inline-block; padding: 2px 10px 2px 10px;'>"#FDDBC7"</span>,<span style='background-color:#F4A582; display: inline-block; padding: 2px 10px 2px 10px;'>"#F4A582"</span>,<span style='background-color:#D6604D; display: inline-block; padding: 2px 10px 2px 10px;'>"#D6604D"</span>,<span style='background-color:#B2182B; display: inline-block; padding: 2px 10px 2px 10px;'>"#B2182B"</span>,<span style='background-color:#67001F; display: inline-block; padding: 2px 10px 2px 10px;'>"#67001F"</span>)</span></div><!--/html_preserve-->
}}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_diverging_BuYlRd
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_diverging_BuYlRd</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 11 colors.</small><br><span style='background-color:#313695; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4575B4; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#74ADD1; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#ABD9E9; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E0F3F8; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FFFFBF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FEE090; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FDAE61; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F46D43; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D73027; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#A50026; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#313695; display: inline-block; padding: 2px 10px 2px 10px;'>"#313695"</span>,<span style='background-color:#4575B4; display: inline-block; padding: 2px 10px 2px 10px;'>"#4575B4"</span>,<span style='background-color:#74ADD1; display: inline-block; padding: 2px 10px 2px 10px;'>"#74ADD1"</span>,<span style='background-color:#ABD9E9; display: inline-block; padding: 2px 10px 2px 10px;'>"#ABD9E9"</span>,<span style='background-color:#E0F3F8; display: inline-block; padding: 2px 10px 2px 10px;'>"#E0F3F8"</span>,<span style='background-color:#FFFFBF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FFFFBF"</span>,<span style='background-color:#FEE090; display: inline-block; padding: 2px 10px 2px 10px;'>"#FEE090"</span>,<span style='background-color:#FDAE61; display: inline-block; padding: 2px 10px 2px 10px;'>"#FDAE61"</span>,<span style='background-color:#F46D43; display: inline-block; padding: 2px 10px 2px 10px;'>"#F46D43"</span>,<span style='background-color:#D73027; display: inline-block; padding: 2px 10px 2px 10px;'>"#D73027"</span>,<span style='background-color:#A50026; display: inline-block; padding: 2px 10px 2px 10px;'>"#A50026"</span>)</span></div><!--/html_preserve-->
}}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_diverging_spectral
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_diverging_spectral</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 96 colors, downsampled to 42 colors.</small><br><span style='background-color:#5b53a4; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#525fa9; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#486cb0; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3f77b5; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3389bd; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3d95b8; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#47a0b3; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#58b2ac; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#64c0a6; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#71c6a5; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#86cfa5; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#94d4a4; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#a2d9a4; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#b5e1a2; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#c3e79f; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#cfec9d; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#e1f399; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#e9f69d; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#eef8a4; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#f6fbb0; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#fcfeba; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#fffdbc; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#fff7b2; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#feeda1; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#fee797; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#fee08b; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#fed27f; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#fdc776; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#fdbd6d; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#fdad60; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#fba05b; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#f99153; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#f67f4b; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#f47044; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#ef6645; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#e55749; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#df4e4b; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#d9444d; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#cb334d; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#be254a; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#b41947; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#a90d45; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#5b53a4; display: inline-block; padding: 2px 10px 2px 10px;'>"#5b53a4"</span>,<span style='background-color:#525fa9; display: inline-block; padding: 2px 10px 2px 10px;'>"#525fa9"</span>,<span style='background-color:#486cb0; display: inline-block; padding: 2px 10px 2px 10px;'>"#486cb0"</span>,<span style='background-color:#3f77b5; display: inline-block; padding: 2px 10px 2px 10px;'>"#3f77b5"</span>,<span style='background-color:#3389bd; display: inline-block; padding: 2px 10px 2px 10px;'>"#3389bd"</span>,<span style='background-color:#3d95b8; display: inline-block; padding: 2px 10px 2px 10px;'>"#3d95b8"</span>,<span style='background-color:#47a0b3; display: inline-block; padding: 2px 10px 2px 10px;'>"#47a0b3"</span>,<span style='background-color:#58b2ac; display: inline-block; padding: 2px 10px 2px 10px;'>"#58b2ac"</span>,<span style='background-color:#64c0a6; display: inline-block; padding: 2px 10px 2px 10px;'>"#64c0a6"</span>,<span style='background-color:#71c6a5; display: inline-block; padding: 2px 10px 2px 10px;'>"#71c6a5"</span>,<span style='background-color:#86cfa5; display: inline-block; padding: 2px 10px 2px 10px;'>"#86cfa5"</span>,<span style='background-color:#94d4a4; display: inline-block; padding: 2px 10px 2px 10px;'>"#94d4a4"</span>,<span style='background-color:#a2d9a4; display: inline-block; padding: 2px 10px 2px 10px;'>"#a2d9a4"</span>,<span style='background-color:#b5e1a2; display: inline-block; padding: 2px 10px 2px 10px;'>"#b5e1a2"</span>,<span style='background-color:#c3e79f; display: inline-block; padding: 2px 10px 2px 10px;'>"#c3e79f"</span>,<span style='background-color:#cfec9d; display: inline-block; padding: 2px 10px 2px 10px;'>"#cfec9d"</span>,<span style='background-color:#e1f399; display: inline-block; padding: 2px 10px 2px 10px;'>"#e1f399"</span>,<span style='background-color:#e9f69d; display: inline-block; padding: 2px 10px 2px 10px;'>"#e9f69d"</span>,<span style='background-color:#eef8a4; display: inline-block; padding: 2px 10px 2px 10px;'>"#eef8a4"</span>,<span style='background-color:#f6fbb0; display: inline-block; padding: 2px 10px 2px 10px;'>"#f6fbb0"</span>,<span style='background-color:#fcfeba; display: inline-block; padding: 2px 10px 2px 10px;'>"#fcfeba"</span>,<span style='background-color:#fffdbc; display: inline-block; padding: 2px 10px 2px 10px;'>"#fffdbc"</span>,<span style='background-color:#fff7b2; display: inline-block; padding: 2px 10px 2px 10px;'>"#fff7b2"</span>,<span style='background-color:#feeda1; display: inline-block; padding: 2px 10px 2px 10px;'>"#feeda1"</span>,<span style='background-color:#fee797; display: inline-block; padding: 2px 10px 2px 10px;'>"#fee797"</span>,<span style='background-color:#fee08b; display: inline-block; padding: 2px 10px 2px 10px;'>"#fee08b"</span>,<span style='background-color:#fed27f; display: inline-block; padding: 2px 10px 2px 10px;'>"#fed27f"</span>,<span style='background-color:#fdc776; display: inline-block; padding: 2px 10px 2px 10px;'>"#fdc776"</span>,<span style='background-color:#fdbd6d; display: inline-block; padding: 2px 10px 2px 10px;'>"#fdbd6d"</span>,<span style='background-color:#fdad60; display: inline-block; padding: 2px 10px 2px 10px;'>"#fdad60"</span>,<span style='background-color:#fba05b; display: inline-block; padding: 2px 10px 2px 10px;'>"#fba05b"</span>,<span style='background-color:#f99153; display: inline-block; padding: 2px 10px 2px 10px;'>"#f99153"</span>,<span style='background-color:#f67f4b; display: inline-block; padding: 2px 10px 2px 10px;'>"#f67f4b"</span>,<span style='background-color:#f47044; display: inline-block; padding: 2px 10px 2px 10px;'>"#f47044"</span>,<span style='background-color:#ef6645; display: inline-block; padding: 2px 10px 2px 10px;'>"#ef6645"</span>,<span style='background-color:#e55749; display: inline-block; padding: 2px 10px 2px 10px;'>"#e55749"</span>,<span style='background-color:#df4e4b; display: inline-block; padding: 2px 10px 2px 10px;'>"#df4e4b"</span>,<span style='background-color:#d9444d; display: inline-block; padding: 2px 10px 2px 10px;'>"#d9444d"</span>,<span style='background-color:#cb334d; display: inline-block; padding: 2px 10px 2px 10px;'>"#cb334d"</span>,<span style='background-color:#be254a; display: inline-block; padding: 2px 10px 2px 10px;'>"#be254a"</span>,<span style='background-color:#b41947; display: inline-block; padding: 2px 10px 2px 10px;'>"#b41947"</span>,<span style='background-color:#a90d45; display: inline-block; padding: 2px 10px 2px 10px;'>"#a90d45"</span>)</span></div><!--/html_preserve-->
}}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_diverging_icefire
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_diverging_icefire</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 96 colors, downsampled to 42 colors.</small><br><span style='background-color:#b7e3d9; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#a9d9d6; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#98cdd2; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#8ac4d0; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#72b6ce; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#63adcd; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#55a3cd; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4394ce; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3987cf; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#377cd0; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3f69c9; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#465ebe; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4954b0; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#474792; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#42407b; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3c3a69; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#323050; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#2c2b42; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#272636; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#212028; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#1f1e21; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#201e1e; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#261e1f; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#332023; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3d2228; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4a252e; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#5c2935; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#6d2b3b; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#7b2d40; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#932e44; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#a22f44; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#b33341; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#c53c3c; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#d24737; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#da5334; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#e66734; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#eb753a; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#ef8445; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#f39a5f; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#f7ab75; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#fab887; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#fcc69a; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#b7e3d9; display: inline-block; padding: 2px 10px 2px 10px;'>"#b7e3d9"</span>,<span style='background-color:#a9d9d6; display: inline-block; padding: 2px 10px 2px 10px;'>"#a9d9d6"</span>,<span style='background-color:#98cdd2; display: inline-block; padding: 2px 10px 2px 10px;'>"#98cdd2"</span>,<span style='background-color:#8ac4d0; display: inline-block; padding: 2px 10px 2px 10px;'>"#8ac4d0"</span>,<span style='background-color:#72b6ce; display: inline-block; padding: 2px 10px 2px 10px;'>"#72b6ce"</span>,<span style='background-color:#63adcd; display: inline-block; padding: 2px 10px 2px 10px;'>"#63adcd"</span>,<span style='background-color:#55a3cd; display: inline-block; padding: 2px 10px 2px 10px;'>"#55a3cd"</span>,<span style='background-color:#4394ce; display: inline-block; padding: 2px 10px 2px 10px;'>"#4394ce"</span>,<span style='background-color:#3987cf; display: inline-block; padding: 2px 10px 2px 10px;'>"#3987cf"</span>,<span style='background-color:#377cd0; display: inline-block; padding: 2px 10px 2px 10px;'>"#377cd0"</span>,<span style='background-color:#3f69c9; display: inline-block; padding: 2px 10px 2px 10px;'>"#3f69c9"</span>,<span style='background-color:#465ebe; display: inline-block; padding: 2px 10px 2px 10px;'>"#465ebe"</span>,<span style='background-color:#4954b0; display: inline-block; padding: 2px 10px 2px 10px;'>"#4954b0"</span>,<span style='background-color:#474792; display: inline-block; padding: 2px 10px 2px 10px;'>"#474792"</span>,<span style='background-color:#42407b; display: inline-block; padding: 2px 10px 2px 10px;'>"#42407b"</span>,<span style='background-color:#3c3a69; display: inline-block; padding: 2px 10px 2px 10px;'>"#3c3a69"</span>,<span style='background-color:#323050; display: inline-block; padding: 2px 10px 2px 10px;'>"#323050"</span>,<span style='background-color:#2c2b42; display: inline-block; padding: 2px 10px 2px 10px;'>"#2c2b42"</span>,<span style='background-color:#272636; display: inline-block; padding: 2px 10px 2px 10px;'>"#272636"</span>,<span style='background-color:#212028; display: inline-block; padding: 2px 10px 2px 10px;'>"#212028"</span>,<span style='background-color:#1f1e21; display: inline-block; padding: 2px 10px 2px 10px;'>"#1f1e21"</span>,<span style='background-color:#201e1e; display: inline-block; padding: 2px 10px 2px 10px;'>"#201e1e"</span>,<span style='background-color:#261e1f; display: inline-block; padding: 2px 10px 2px 10px;'>"#261e1f"</span>,<span style='background-color:#332023; display: inline-block; padding: 2px 10px 2px 10px;'>"#332023"</span>,<span style='background-color:#3d2228; display: inline-block; padding: 2px 10px 2px 10px;'>"#3d2228"</span>,<span style='background-color:#4a252e; display: inline-block; padding: 2px 10px 2px 10px;'>"#4a252e"</span>,<span style='background-color:#5c2935; display: inline-block; padding: 2px 10px 2px 10px;'>"#5c2935"</span>,<span style='background-color:#6d2b3b; display: inline-block; padding: 2px 10px 2px 10px;'>"#6d2b3b"</span>,<span style='background-color:#7b2d40; display: inline-block; padding: 2px 10px 2px 10px;'>"#7b2d40"</span>,<span style='background-color:#932e44; display: inline-block; padding: 2px 10px 2px 10px;'>"#932e44"</span>,<span style='background-color:#a22f44; display: inline-block; padding: 2px 10px 2px 10px;'>"#a22f44"</span>,<span style='background-color:#b33341; display: inline-block; padding: 2px 10px 2px 10px;'>"#b33341"</span>,<span style='background-color:#c53c3c; display: inline-block; padding: 2px 10px 2px 10px;'>"#c53c3c"</span>,<span style='background-color:#d24737; display: inline-block; padding: 2px 10px 2px 10px;'>"#d24737"</span>,<span style='background-color:#da5334; display: inline-block; padding: 2px 10px 2px 10px;'>"#da5334"</span>,<span style='background-color:#e66734; display: inline-block; padding: 2px 10px 2px 10px;'>"#e66734"</span>,<span style='background-color:#eb753a; display: inline-block; padding: 2px 10px 2px 10px;'>"#eb753a"</span>,<span style='background-color:#ef8445; display: inline-block; padding: 2px 10px 2px 10px;'>"#ef8445"</span>,<span style='background-color:#f39a5f; display: inline-block; padding: 2px 10px 2px 10px;'>"#f39a5f"</span>,<span style='background-color:#f7ab75; display: inline-block; padding: 2px 10px 2px 10px;'>"#f7ab75"</span>,<span style='background-color:#fab887; display: inline-block; padding: 2px 10px 2px 10px;'>"#fab887"</span>,<span style='background-color:#fcc69a; display: inline-block; padding: 2px 10px 2px 10px;'>"#fcc69a"</span>)</span></div><!--/html_preserve-->
}}
}
\keyword{datasets}
