% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dinosaurs}
\alias{dinosaurs}
\title{Dinosaurs data}
\format{
A data frame with 800 rows and 9 variables.
}
\source{
\code{dinosaurs} package
}
\usage{
dinosaurs
}
\description{
Dinosaurs data
}
\keyword{datasets}
