% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disease_course}
\alias{disease_course}
\title{Disease course data}
\format{
A data frame with 800 rows and 9 variables.
}
\source{
\code{tidyplots} package
}
\usage{
disease_course
}
\description{
Disease course data
}
\keyword{datasets}
