% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distributions}
\alias{distributions}
\title{Distributions data}
\format{
A data frame with 800 rows and 9 variables.
}
\source{
\code{tidyplots} package
}
\usage{
distributions
}
\description{
Distributions data
}
\keyword{datasets}
