% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{energy}
\alias{energy}
\title{Energy data}
\format{
A data frame with 242 rows and 4 variables.
}
\source{
\href{https://www.energy-charts.info}{Energy-Charts}
}
\usage{
energy
}
\description{
Energy data
}
\keyword{datasets}
