% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{energy_week}
\alias{energy_week}
\title{Energy week data}
\format{
A data frame with 10,080 rows and 4 variables.
}
\source{
\href{https://www.energy-charts.info}{Energy-Charts}
}
\usage{
energy_week
}
\description{
Energy week data
}
\keyword{datasets}
