% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eu_countries}
\alias{eu_countries}
\title{EU countries data}
\format{
A data frame with 800 rows and 9 variables.
}
\source{
\code{eu_countries} package
}
\usage{
eu_countries
}
\description{
EU countries data
}
\keyword{datasets}
