% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{format_p_value}
\alias{format_p_value}
\title{Format p values}
\usage{
format_p_value(x, accuracy = 1e-04)
}
\arguments{
\item{x}{A \code{number} to format.}

\item{accuracy}{A number to round to. For example, use \code{0.01} to show 2 decimal
places of precision. Defaults to \code{0.0001}, corresponding to 4 decimal
places of precision.}
}
\value{
Formatted number as \code{character} string.
}
\description{
Format p values
}
\examples{
format_p_value(0.03445553)
format_p_value(0.0003445553)
format_p_value(0.00003445553)

}
