% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_statistics}
\alias{gene_statistics}
\title{RNA-Seq differential expression data}
\format{
A data frame with 800 rows and 9 variables.
}
\source{
\code{tidyplots} package
}
\usage{
gene_statistics
}
\description{
RNA-Seq differential expression data
}
\keyword{datasets}
