% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{general_arguments}
\alias{general_arguments}
\title{General arguments}
\arguments{
\item{plot}{A tidyplot generated with the \code{tidyplot()} function.}

\item{data}{A function to subset the data to be plotted. See \code{filter_rows()} and friends.}

\item{dodge_width}{For adjusting the distance between grouped objects.}

\item{preserve}{general}

\item{rasterize}{general}

\item{rasterize_dpi}{general}
}
\description{
General arguments
}
