% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%>\%}
\alias{\%>\%}
\title{The pipe}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{A value.}

\item{rhs}{A function call.}
}
\value{
The result of the calling the function \code{rhs} with the parameter \code{lhs}.
}
\description{
The pipe
}
\keyword{internal}
