% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pca}
\alias{pca}
\title{Principle component analysis data}
\format{
A data frame.
}
\source{
\href{https://www.nature.com/articles/s41593-019-0385-4}{Bassoon proteinopathy drives neurodegeneration in multiple sclerosis}, Nature Neuroscience 2019
}
\usage{
pca
}
\description{
Principle component analysis data
}
\examples{
dplyr::glimpse(pca)
}
\keyword{datasets}
