% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_clipping}
\alias{remove_clipping}
\title{Remove plot area clipping}
\usage{
remove_clipping(plot)
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}
}
\value{
A \code{tidyplot} object.
}
\description{
Remove plot area clipping
}
\examples{
# Before removing
spendings |>
  tidyplot(x = amount, y = category, color = category) |>
  add_sum_bar(alpha = 0.2) |>
  add_sum_dash() |>
  add_sum_value(accuracy = 1, color = "black")

# After removing
spendings |>
  tidyplot(x = amount, y = category, color = category) |>
  add_sum_bar(alpha = 0.2) |>
  add_sum_dash() |>
  add_sum_value(accuracy = 1, color = "black") |>
  remove_clipping()

}
