% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_padding}
\alias{remove_padding}
\title{Remove padding}
\usage{
remove_padding(gg, force_y_continuous = FALSE)
}
\arguments{
\item{gg}{bla}

\item{force_y_continuous}{bla}
}
\description{
Remove padding
}
