% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_plot_area_padding}
\alias{remove_plot_area_padding}
\title{Remove plot area padding}
\usage{
remove_plot_area_padding(plot, force_continuous = FALSE)
}
\arguments{
\item{plot}{bla}

\item{force_continuous}{bla}
}
\description{
Remove plot area padding
}
