% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_x_axis}
\alias{remove_x_axis}
\alias{remove_x_axis_line}
\alias{remove_x_axis_ticks}
\alias{remove_x_axis_labels}
\alias{remove_x_axis_title}
\title{Remove x axis or parts of it}
\usage{
remove_x_axis(gg)

remove_x_axis_line(gg)

remove_x_axis_ticks(gg)

remove_x_axis_labels(gg)

remove_x_axis_title(gg)
}
\arguments{
\item{gg}{bla}
}
\description{
Remove x axis or parts of it
}
