% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_y_axis}
\alias{remove_y_axis}
\alias{remove_y_axis_line}
\alias{remove_y_axis_ticks}
\alias{remove_y_axis_labels}
\alias{remove_y_axis_title}
\title{Remove y axis or parts of it}
\usage{
remove_y_axis(plot)

remove_y_axis_line(plot)

remove_y_axis_ticks(plot)

remove_y_axis_labels(plot)

remove_y_axis_title(plot)
}
\arguments{
\item{plot}{bla}
}
\description{
Remove y axis or parts of it
}
