% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{rename_x_axis_labels}
\alias{rename_x_axis_labels}
\alias{rename_y_axis_labels}
\alias{rename_color_labels}
\title{Rename axis or color labels}
\usage{
rename_x_axis_labels(plot, new_names)

rename_y_axis_labels(plot, new_names)

rename_color_labels(plot, new_names)
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}

\item{new_names}{Named character vector in the format c("old1" = "new1", "old2" = "new2").}
}
\value{
A \code{tidyplot} object
}
\description{
Rename axis or color labels
}
\examples{
study \%>\%
  tidyplot(x = treatment, y = score, color = group) \%>\%
  add_data_points() \%>\%
  add_mean_bar(alpha = 0.3) \%>\%
  add_sem_bar() \%>\%
  rename_x_axis_labels(new_names = c(
    "A" = "This",
    "B" = "is",
    "C" = "totally",
    "D" = "new"))
study \%>\%
  tidyplot(x = score, y = treatment, color = group) \%>\%
  add_data_points() \%>\%
  add_mean_bar(alpha = 0.3) \%>\%
  add_sem_bar() \%>\%
  rename_y_axis_labels(new_names = c(
    "A" = "This",
    "B" = "is",
    "C" = "totally",
    "D" = "new"))
study \%>\%
  tidyplot(x = treatment, y = score, color = group) \%>\%
  add_data_points() \%>\%
  add_mean_bar(alpha = 0.3) \%>\%
  add_sem_bar() \%>\%
  rename_color_labels(new_names = c(
    "placebo" = "The first",
    "treatment" = "The second"))

}
