% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{render_plot}
\alias{render_plot}
\title{Render plot}
\usage{
render_plot(gg, data = all_rows(), title = ggplot2::waiver(), ...)
}
\arguments{
\item{gg}{bla}

\item{data}{bla}

\item{title}{bla}

\item{...}{bla}
}
\description{
Render plot
}
