% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{reverse_x_axis_labels}
\alias{reverse_x_axis_labels}
\alias{reverse_y_axis_labels}
\alias{reverse_color_labels}
\title{Reverse axis or color labels}
\usage{
reverse_x_axis_labels(plot)

reverse_y_axis_labels(plot)

reverse_color_labels(plot)
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}
}
\value{
A \code{tidyplot} object
}
\description{
Reverse axis or color labels
}
\examples{
study \%>\%
  tidyplot(x = treatment, y = score, color = group) \%>\%
  add_data_points() \%>\%
  add_mean_bar(alpha = 0.3) \%>\%
  add_sem_bar() \%>\%
  reverse_x_axis_labels()
study \%>\%
  tidyplot(x = score, y = treatment, color = group) \%>\%
  add_data_points() \%>\%
  add_mean_bar(alpha = 0.3) \%>\%
  add_sem_bar() \%>\%
  reverse_y_axis_labels()
study \%>\%
  tidyplot(x = treatment, y = score, color = group) \%>\%
  add_data_points() \%>\%
  add_mean_bar(alpha = 0.3) \%>\%
  add_sem_bar() \%>\%
  reverse_color_labels()

}
