% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{save_plot}
\alias{save_plot}
\title{Save plots to file}
\usage{
save_plot(
  plot = ggplot2::last_plot(),
  filename,
  device = NULL,
  path = NULL,
  scale = 1,
  width = NA,
  height = NA,
  units = c("mm", "cm", "in"),
  dpi = 300,
  limitsize = TRUE,
  multiple_files = FALSE,
  bg = "transparent",
  ...
)
}
\arguments{
\item{plot}{Plot to save, defaults to last plot displayed.}

\item{filename}{File name to create on disk.}

\item{device}{Device to use. Can either be a device function
(e.g. \link{png}), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only). If
\code{NULL} (default), the device is guessed based on the \code{filename} extension.}

\item{path}{Path of the directory to save plot to: \code{path} and \code{filename}
are combined to create the fully qualified file name. Defaults to the
working directory.}

\item{scale}{Multiplicative scaling factor.}

\item{width, height}{Dimensions of the saved plot. Defaults to \code{NA}.}

\item{units}{Units of length. Defaults to \code{"mm"}.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{limitsize}{When \code{TRUE} (the default), \code{ggsave()} will not
save images larger than 50x50 inches, to prevent the common error of
specifying dimensions in pixels.}

\item{multiple_files}{Whether to save multiple pages as individual files.}

\item{bg}{Background colour. If \code{NULL}, uses the \code{plot.background} fill value
from the plot theme.}

\item{...}{Other arguments passed on to the graphics device function,
as specified by \code{device}.}
}
\value{
A \code{tidyplot} object
}
\description{
This function takes a plot or list of plots and writes them to a (multipage) file.
}
\details{
\strong{Handling of file dimensions.}
Output file dimensions are determined according the the following precedence.
\enumerate{
\item The \code{width} and \code{height} arguments.
\item Dimensions inferred from an incoming \code{plot} object with absolute dimensions.
\item System default device dimensions.
}
}
