% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{show_color_palettes}
\alias{show_color_palettes}
\title{Show color palettes}
\usage{
show_color_palettes(pals = color_palettes, show_labels = TRUE, ncol = 3, ...)
}
\arguments{
\item{pals}{bla}

\item{show_labels}{bla}

\item{ncol}{bla}

\item{...}{bla}
}
\description{
Show color palettes
}
