% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{show_plot}
\alias{show_plot}
\title{Show plot on screen}
\usage{
show_plot(gg, data = all_rows(), title = ggplot2::waiver(), ...)
}
\arguments{
\item{gg}{bla}

\item{data}{bla}

\item{title}{bla}

\item{...}{bla}
}
\description{
Show plot on screen
}
