% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spendings}
\alias{spendings}
\title{Spending data}
\format{
A data frame with 800 rows and 9 variables.
}
\source{
\code{tidyplots} package
}
\usage{
spendings
}
\description{
Spending data
}
\keyword{datasets}
