% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_tidyplot}
\alias{theme_tidyplot}
\alias{theme_ggplot2}
\alias{theme_minimal_xy}
\alias{theme_minimal_y}
\alias{theme_minimal_x}
\title{Themes}
\usage{
theme_tidyplot(plot, fontsize = 7)

theme_ggplot2(plot, fontsize = 7)

theme_minimal_xy(plot, fontsize = 7)

theme_minimal_y(plot, fontsize = 7)

theme_minimal_x(plot, fontsize = 7)
}
\arguments{
\item{plot}{bla}

\item{fontsize}{bla}
}
\description{
Themes
}
