% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time_course}
\alias{time_course}
\title{Time course data}
\format{
A data frame with 800 rows and 9 variables.
}
\source{
\code{tidyplots} package
}
\usage{
time_course
}
\description{
Time course data
}
\keyword{datasets}
