% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{view_plot}
\alias{view_plot}
\title{View plot on screen}
\usage{
view_plot(plot, data = all_rows(), title = ggplot2::waiver(), ...)
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}

\item{data}{The data to be displayed in this layer. There are three options:
\itemize{
\item If \code{all_rows()} (the default) the complete dataset is displayed.
\item A \code{function} to subset the plot data. See \code{filter_rows()} and friends.
\item A \code{data.frame} to override the plot data.
}}

\item{title}{Plot title.}

\item{...}{Arguments passed on to \code{print()}.}
}
\value{
A \code{tidyplot} object
}
\description{
View plot on screen
}
