% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add \code{ggplot2} code}
\usage{
add()
}
\description{
Add \code{ggplot2} code
}
\examples{
df_demo \%>\%
  tidyplot(x = category, y = value, color = category) \%>\%
  add(ggplot2::geom_point())

}
