% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-misc.R
\name{add}
\alias{add}
\title{Add ggplot2 code to a tidyplot}
\usage{
add()
}
\value{
A \code{tidyplot} object.
}
\description{
Add ggplot2 code to a tidyplot
}
\examples{
study \%>\%
  tidyplot(x = treatment, y = score, color = treatment) \%>\%
  add(ggplot2::geom_point())

}
