% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-proportional.R
\name{add_areastack_absolute}
\alias{add_areastack_absolute}
\alias{add_areastack_relative}
\title{Add area stack}
\usage{
add_areastack_absolute(
  plot,
  linewidth = 0.25,
  alpha = 0.3,
  reverse = FALSE,
  replace_na = FALSE,
  ...
)

add_areastack_relative(
  plot,
  linewidth = 0.25,
  alpha = 0.3,
  reverse = FALSE,
  replace_na = FALSE,
  ...
)
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}

\item{linewidth}{Thickness of the line in points (pt). Typical values range between \code{0.25} and \code{1}.}

\item{alpha}{A \code{number} between \code{0} and \code{1} for the transparency of an object. A value of \code{0} is completely transparent, \code{1} is completely opaque.}

\item{reverse}{Whether the order should be reversed or not. Defaults to \code{FALSE}, meaning not reversed.}

\item{replace_na}{Whether to replace \code{count = NA} with \code{count = 0}.}

\item{...}{Arguments passed on to the \code{geom} function.}
}
\value{
A \code{tidyplot} object
}
\description{
Add area stack
}
