% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-proportional.R
\name{add_areastack_absolute}
\alias{add_areastack_absolute}
\alias{add_areastack_relative}
\title{Add area stack}
\usage{
add_areastack_absolute(
  plot,
  linewidth = 0.25,
  alpha = 0.3,
  reverse = FALSE,
  replace_na = FALSE,
  ...
)

add_areastack_relative(
  plot,
  linewidth = 0.25,
  alpha = 0.3,
  reverse = FALSE,
  replace_na = FALSE,
  ...
)
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}

\item{linewidth}{Thickness of the line in points (pt). Typical values range between \code{0.25} and \code{1}.}

\item{alpha}{common}

\item{reverse}{common}

\item{replace_na}{common}

\item{...}{Arguments passed on to the \code{geom} function.}
}
\value{
A \code{tidyplot} object
}
\description{
Add area stack
}
