% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_barstack_absolute}
\alias{add_barstack_absolute}
\alias{add_barstack_relative}
\title{Add bar stack}
\usage{
add_barstack_absolute(plot, width = 0.8, reverse = FALSE, ...)

add_barstack_relative(plot, width = 0.8, reverse = FALSE, ...)
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}

\item{width}{Horizontal width of the plotted object (bar, error bar, boxplot,
violin plot, etc). Typical values range between \code{0} and \code{1}.}

\item{reverse}{common}

\item{...}{Arguments passed on to the \code{geom} function.}
}
\value{
A \code{tidyplot} object
}
\description{
Add bar stack
}
