% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_barstack_absolute}
\alias{add_barstack_absolute}
\alias{add_barstack_relative}
\title{Add bar stack}
\usage{
add_barstack_absolute(gg, bar_width = 0.8, reverse = FALSE, ...)

add_barstack_relative(gg, bar_width = 0.8, reverse = FALSE, ...)
}
\arguments{
\item{gg}{bla}

\item{bar_width}{bla}

\item{reverse}{bla}

\item{...}{bla}
}
\description{
Add bar stack
}
