% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_count_bar}
\alias{add_count_bar}
\alias{add_count_dash}
\alias{add_count_dot}
\alias{add_count_value}
\alias{add_count_line}
\alias{add_count_area}
\title{Add count}
\usage{
add_count_bar(
  plot,
  dodge_width = NULL,
  width = 0.6,
  saturation = 1,
  preserve = "total",
  ...
)

add_count_dash(
  plot,
  dodge_width = NULL,
  width = 0.6,
  linewidth = 0.25,
  preserve = "total",
  ...
)

add_count_dot(plot, dodge_width = NULL, size = 2, preserve = "total", ...)

add_count_value(
  plot,
  dodge_width = NULL,
  accuracy = 0.1,
  scale_cut = NULL,
  fontsize = 7,
  vjust = -0.5,
  padding_top = 0.15,
  preserve = "total",
  ...
)

add_count_line(
  plot,
  group,
  dodge_width = NULL,
  linewidth = 0.25,
  preserve = "total",
  ...
)

add_count_area(
  plot,
  group,
  dodge_width = NULL,
  linewidth = 0.25,
  preserve = "total",
  ...
)
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}

\item{dodge_width}{For adjusting the distance between grouped objects.}

\item{width}{Horizontal width of the plotted object (bar, error bar, boxplot,
violin plot, etc). Typical values range between \code{0} and \code{1}.}

\item{saturation}{common}

\item{preserve}{Should dodging preserve the \code{"total"} width of all elements at
a position, or the width of a \code{"single"} element?}

\item{...}{Arguments passed on to the \code{geom} function.}

\item{linewidth}{Thickness of the line in points (pt). Typical values range between \code{0.25} and \code{1}.}

\item{size}{A \code{number} representing the size of the plot symbol. Typical
values range between \code{1} and \code{3}.}

\item{accuracy}{A number to round to. Use (e.g.) \code{0.01} to show 2 decimal
places of precision. If \code{NULL}, the default, uses a heuristic that should
ensure breaks have the minimum number of digits needed to show the
difference between adjacent values.

Applied to rescaled data.}

\item{scale_cut}{common}

\item{fontsize}{common}

\item{vjust}{bla}

\item{padding_top}{bla}

\item{group}{common}
}
\value{
A \code{tidyplot} object
}
\description{
Add count
}
