% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_count_bar}
\alias{add_count_bar}
\alias{add_count_dash}
\alias{add_count_dot}
\alias{add_count_value}
\alias{add_count_line}
\title{Add count}
\usage{
add_count_bar(gg, dodge_width = NULL, alpha = 1, width = 0.6, ...)

add_count_dash(gg, dodge_width = NULL, width = 0.6, linewidth = 0.25, ...)

add_count_dot(gg, dodge_width = NULL, size = 2, ...)

add_count_value(
  gg,
  accuracy = 0.1,
  scale_cut = NULL,
  dodge_width = NULL,
  fontsize = 7,
  vjust = -0.5,
  padding_top = 0.15,
  ...
)

add_count_line(gg, group, dodge_width = NULL, linewidth = 0.25, ...)
}
\arguments{
\item{gg}{bla}

\item{dodge_width}{bla}

\item{alpha}{bla}

\item{width}{bla}

\item{...}{bla}

\item{linewidth}{bla}

\item{size}{bla}

\item{accuracy}{bla}

\item{scale_cut}{bla}

\item{fontsize}{bla}

\item{vjust}{bla}

\item{padding_top}{bla}

\item{group}{bla}
}
\description{
Add count
}
