% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_curve}
\alias{add_curve}
\title{Add curve}
\usage{
add_curve(
  gg,
  dodge_width = NULL,
  method = "loess",
  linewidth = 0.25,
  alpha = 0.3,
  ...
)
}
\arguments{
\item{gg}{bla}

\item{dodge_width}{bla}

\item{method}{bla}

\item{linewidth}{bla}

\item{alpha}{bla}

\item{...}{bla}
}
\description{
Add curve
}
