% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_data_points}
\alias{add_data_points}
\alias{add_data_points_jitter}
\alias{add_data_points_beeswarm}
\title{Add data points}
\usage{
add_data_points(
  gg,
  data = all_rows(),
  shape = 19,
  size = 1,
  confetti = FALSE,
  dodge_width = NULL,
  preserve = "total",
  rasterize = FALSE,
  rasterize_dpi = 300,
  ...
)

add_data_points_jitter(
  gg,
  data = all_rows(),
  shape = 19,
  size = 1,
  confetti = FALSE,
  dodge_width = NULL,
  jitter_width = 0.2,
  jitter_height = 0,
  preserve = "total",
  rasterize = FALSE,
  rasterize_dpi = 300,
  ...
)

add_data_points_beeswarm(
  gg,
  data = all_rows(),
  shape = 19,
  size = 1,
  confetti = FALSE,
  cex = 3,
  corral = "wrap",
  corral.width = 0.5,
  dodge_width = NULL,
  preserve = "total",
  rasterize = FALSE,
  rasterize_dpi = 300,
  ...
)
}
\arguments{
\item{gg}{bla}

\item{data}{bla}

\item{shape}{bla}

\item{size}{bla}

\item{confetti}{bla}

\item{dodge_width}{bla}

\item{preserve}{bla}

\item{rasterize}{bla}

\item{rasterize_dpi}{bla}

\item{...}{bla}

\item{jitter_width}{bla}

\item{jitter_height}{bla}

\item{cex}{Scaling for adjusting point spacing (see \code{\link[beeswarm:swarmx]{beeswarm::swarmx()}}).
Values between 1 (default) and 3 tend to work best.}

\item{corral}{\code{string}. Method used to adjust points that would be placed to
wide horizontally, default is \code{"none"}. See details below.}

\item{corral.width}{\code{numeric}. Width of the corral, default is \code{0.9}.}
}
\description{
Add data points
}
\examples{
study \%>\%
  tidyplot(x = treatment, y = score, color = treatment) \%>\%
  add_mean_bar(alpha = 0.3) \%>\%
  add_error() \%>\%
  add_data_points_beeswarm()

}
