% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_error}
\alias{add_error}
\alias{add_range}
\alias{add_sd}
\alias{add_ci95}
\title{Add error bar}
\usage{
add_error(gg, dodge_width = NULL, width = 0.4, linewidth = 0.25, ...)

add_range(gg, dodge_width = NULL, width = 0.4, linewidth = 0.25, ...)

add_sd(gg, dodge_width = NULL, width = 0.4, linewidth = 0.25, ...)

add_ci95(gg, dodge_width = NULL, width = 0.4, linewidth = 0.25, ...)
}
\arguments{
\item{gg}{bla}

\item{dodge_width}{bla}

\item{width}{bla}

\item{linewidth}{bla}

\item{...}{bla}
}
\description{
Add error bar
}
