% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_error}
\alias{add_error}
\alias{add_range}
\alias{add_sd}
\alias{add_ci95}
\title{Add error bar}
\usage{
add_error(
  plot,
  dodge_width = NULL,
  width = 0.4,
  linewidth = 0.25,
  preserve = "total",
  ...
)

add_range(
  plot,
  dodge_width = NULL,
  width = 0.4,
  linewidth = 0.25,
  preserve = "total",
  ...
)

add_sd(
  plot,
  dodge_width = NULL,
  width = 0.4,
  linewidth = 0.25,
  preserve = "total",
  ...
)

add_ci95(
  plot,
  dodge_width = NULL,
  width = 0.4,
  linewidth = 0.25,
  preserve = "total",
  ...
)
}
\arguments{
\item{plot}{A tidyplot generated with the \code{tidyplot()} function.}

\item{dodge_width}{For adjusting the distance between grouped objects.}

\item{width}{bla}

\item{linewidth}{bla}

\item{preserve}{general}

\item{...}{bla}
}
\description{
\itemize{
\item \code{add_error()} adds the standard error of mean.
\item \code{add_range()} adds the range from smallest to largest value.
\item \code{add_sd()} adds the standard deviation.
\item \code{add_ci95()} adds the 95\% confidence interval.
}
}
