% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_error_ribbon}
\alias{add_error_ribbon}
\alias{add_range_ribbon}
\alias{add_sd_ribbon}
\alias{add_ci95_ribbon}
\title{Add ribbon}
\usage{
add_error_ribbon(plot, dodge_width = NULL, alpha = 0.3, color = NA, ...)

add_range_ribbon(plot, dodge_width = NULL, alpha = 0.3, color = NA, ...)

add_sd_ribbon(plot, dodge_width = NULL, alpha = 0.3, color = NA, ...)

add_ci95_ribbon(plot, dodge_width = NULL, alpha = 0.3, color = NA, ...)
}
\arguments{
\item{plot}{A tidyplot generated with the \code{tidyplot()} function.}

\item{dodge_width}{For adjusting the distance between grouped objects.}

\item{alpha}{bla}

\item{color}{bla}

\item{...}{bla}
}
\description{
\itemize{
\item \code{add_error_ribbon()} adds the standard error of mean.
\item \code{add_range_ribbon()} adds the range from smallest to largest value.
\item \code{add_sd_ribbon()} adds the standard deviation.
\item \code{add_ci95_ribbon()} adds the 95\% confidence interval.
}
}
