% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-general.R
\name{add_error_ribbon}
\alias{add_error_ribbon}
\alias{add_range_ribbon}
\alias{add_sd_ribbon}
\alias{add_ci95_ribbon}
\title{Add ribbon}
\usage{
add_error_ribbon(plot, dodge_width = NULL, alpha = 0.3, color = NA, ...)

add_range_ribbon(plot, dodge_width = NULL, alpha = 0.3, color = NA, ...)

add_sd_ribbon(plot, dodge_width = NULL, alpha = 0.3, color = NA, ...)

add_ci95_ribbon(plot, dodge_width = NULL, alpha = 0.3, color = NA, ...)
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}

\item{dodge_width}{For adjusting the distance between grouped objects.}

\item{alpha}{common}

\item{color}{common}

\item{...}{Arguments passed on to the \code{geom} function.}
}
\value{
A \code{tidyplot} object
}
\description{
\itemize{
\item \code{add_error_ribbon()} adds the standard error of mean.
\item \code{add_range_ribbon()} adds the range from smallest to largest value.
\item \code{add_sd_ribbon()} adds the standard deviation.
\item \code{add_ci95_ribbon()} adds the 95\% confidence interval.
}
}
\examples{
time_course \%>\%
  tidyplot(x = day, y = score, color = treatment, dodge_width = 0) \%>\%
  add_mean_line() \%>\%
  add_error_ribbon()
time_course \%>\%
  tidyplot(x = day, y = score, color = treatment, dodge_width = 0) \%>\%
  add_mean_line() \%>\%
  add_range_ribbon()
time_course \%>\%
  tidyplot(x = day, y = score, color = treatment, dodge_width = 0) \%>\%
  add_mean_line() \%>\%
  add_sd_ribbon()
time_course \%>\%
  tidyplot(x = day, y = score, color = treatment, dodge_width = 0) \%>\%
  add_mean_line() \%>\%
  add_ci95_ribbon()

}
