% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_heatmap}
\alias{add_heatmap}
\title{Add heatmap}
\usage{
add_heatmap(
  gg,
  scale = c("none", "row", "column"),
  rotate_labels = 90,
  rasterize = FALSE,
  rasterize_dpi = 300,
  ...
)
}
\arguments{
\item{gg}{bla}

\item{scale}{bla}

\item{rotate_labels}{bla}

\item{rasterize}{bla}

\item{rasterize_dpi}{bla}

\item{...}{bla}
}
\description{
Add heatmap
}
