% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_heatmap}
\alias{add_heatmap}
\title{Add heatmap}
\usage{
add_heatmap(gg, scale = c("none", "row", "column"), ...)
}
\arguments{
\item{gg}{bla}

\item{scale}{bla}

\item{...}{bla}
}
\description{
Add heatmap
}
