% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-misc.R
\name{add_histogram}
\alias{add_histogram}
\alias{add_density_histogram}
\alias{add_density_curve}
\title{Add histogram or density}
\usage{
add_histogram(plot, binwidth = NULL, bins = NULL, ...)

add_density_histogram(plot, binwidth = NULL, bins = NULL, ...)

add_density_curve(
  plot,
  bw = "nrd0",
  adjust = 1,
  kernel = "gaussian",
  n = 512,
  alpha = 0.4,
  color = "#D55E00",
  ...
)
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}

\item{binwidth}{The width of the bins. Can be specified as a numeric value
or as a function that calculates width from unscaled x. Here, "unscaled x"
refers to the original x values in the data, before application of any
scale transformation. When specifying a function along with a grouping
structure, the function will be called once per group.
The default is to use the number of bins in \code{bins},
covering the range of the data. You should always override
this value, exploring multiple widths to find the best to illustrate the
stories in your data.

The bin width of a date variable is the number of days in each time; the
bin width of a time variable is the number of seconds.}

\item{bins}{Number of bins. Overridden by \code{binwidth}. Defaults to 30.}

\item{...}{Arguments passed on to the \code{geom} function.}

\item{bw}{The smoothing bandwidth to be used.
If numeric, the standard deviation of the smoothing kernel.
If character, a rule to choose the bandwidth, as listed in
\code{\link[stats:bandwidth]{stats::bw.nrd()}}. Note that automatic calculation of the bandwidth does
not take weights into account.}

\item{adjust}{A multiplicate bandwidth adjustment. This makes it possible
to adjust the bandwidth while still using the a bandwidth estimator.
For example, \code{adjust = 1/2} means use half of the default bandwidth.}

\item{kernel}{Kernel. See list of available kernels in \code{\link[=density]{density()}}.}

\item{n}{number of equally spaced points at which the density is to be
estimated, should be a power of two, see \code{\link[=density]{density()}} for
details}

\item{alpha}{A \code{number} between \code{0} and \code{1} for the opacity of an object. A value of \code{0} is completely transparent, \code{1} is completely opaque.}

\item{color}{A hex color for the stroke color. For example, \code{"#FFFFFF"} for white.}
}
\value{
A \code{tidyplot} object
}
\description{
Add histogram or density
}
\examples{
energy \%>\%
  tidyplot(x = power) \%>\%
  add_histogram()

energy \%>\%
  tidyplot(x = power) \%>\%
   add_density_histogram() \%>\%
   add_density_curve()

}
