% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_histogram}
\alias{add_histogram}
\alias{add_density_histogram}
\title{Add histogram}
\usage{
add_histogram(gg, binwidth = NULL, bins = NULL, color = "#4DACD6", ...)

add_density_histogram(gg, binwidth = NULL, bins = NULL, color = "#4DACD6", ...)
}
\arguments{
\item{gg}{bla}

\item{binwidth}{bla}

\item{bins}{bla}

\item{color}{bla}

\item{...}{bla}
}
\description{
Add histogram
}
