% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_line}
\alias{add_line}
\alias{add_area}
\title{Add line or area}
\usage{
add_line(
  plot,
  group,
  dodge_width = NULL,
  linewidth = 0.25,
  preserve = "total",
  ...
)

add_area(
  plot,
  group,
  dodge_width = NULL,
  linewidth = 0.25,
  alpha = 0.3,
  preserve = "total",
  ...
)
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}

\item{group}{common}

\item{dodge_width}{For adjusting the distance between grouped objects.}

\item{linewidth}{Thickness of the line in points (pt). Typical values range between \code{0.25} and \code{1}.}

\item{preserve}{Should dodging preserve the \code{"total"} width of all elements at
a position, or the width of a \code{"single"} element?}

\item{...}{Arguments passed on to the \code{geom} function.}

\item{alpha}{common}
}
\value{
A \code{tidyplot} object
}
\description{
Add line or area
}
