% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_line}
\alias{add_line}
\alias{add_area}
\title{Add line}
\usage{
add_line(
  plot,
  group,
  dodge_width = NULL,
  linewidth = 0.25,
  preserve = "total",
  ...
)

add_area(
  plot,
  group,
  dodge_width = NULL,
  linewidth = 0.25,
  preserve = "total",
  ...
)
}
\arguments{
\item{plot}{A tidyplot generated with the \code{tidyplot()} function.}

\item{group}{bla}

\item{dodge_width}{For adjusting the distance between grouped objects.}

\item{linewidth}{bla}

\item{preserve}{general}

\item{...}{bla}
}
\description{
Add line
}
