% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_line}
\alias{add_line}
\alias{add_area}
\title{Add line}
\usage{
add_line(
  gg,
  group,
  dodge_width = NULL,
  linewidth = 0.25,
  preserve = "total",
  ...
)

add_area(gg, dodge_width = NULL, linewidth = 0.25, preserve = "total", ...)
}
\arguments{
\item{gg}{bla}

\item{group}{bla}

\item{dodge_width}{bla}

\item{linewidth}{bla}

\item{preserve}{bla}

\item{...}{bla}
}
\description{
Add line
}
