% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_mean_bar}
\alias{add_mean_bar}
\alias{add_mean_dash}
\alias{add_mean_dot}
\alias{add_mean_value}
\alias{add_mean_line}
\title{Add mean}
\usage{
add_mean_bar(gg, dodge_width = NULL, width = 0.6, alpha = 1, ...)

add_mean_dash(gg, dodge_width = NULL, width = 0.6, linewidth = 0.25, ...)

add_mean_dot(gg, dodge_width = NULL, size = 2, ...)

add_mean_value(
  gg,
  dodge_width = NULL,
  accuracy = 0.1,
  scale_cut = NULL,
  fontsize = 7,
  vjust = -0.5,
  padding_top = 0.15,
  ...
)

add_mean_line(gg, group, dodge_width = NULL, linewidth = 0.25, ...)
}
\arguments{
\item{gg}{bla}

\item{dodge_width}{bla}

\item{width}{bla}

\item{alpha}{description}

\item{...}{description}

\item{linewidth}{bla}

\item{size}{description}

\item{accuracy}{description}

\item{scale_cut}{bla}

\item{fontsize}{description}

\item{vjust}{description}

\item{padding_top}{description}

\item{group}{bla}
}
\description{
Add mean
}
