% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_pie}
\alias{add_pie}
\alias{add_donut}
\title{Add pie chart}
\usage{
add_pie(gg, bar_width = 1, reverse = FALSE, ...)

add_donut(gg, bar_width = 1, reverse = FALSE, ...)
}
\arguments{
\item{gg}{bla}

\item{bar_width}{bla}

\item{reverse}{bla}

\item{...}{bla}
}
\description{
Add pie chart
}
