% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_pie}
\alias{add_pie}
\alias{add_donut}
\title{Add pie or donut chart}
\usage{
add_pie(plot, width = 1, reverse = FALSE, ...)

add_donut(plot, width = 1, reverse = FALSE, ...)
}
\arguments{
\item{plot}{A tidyplot generated with the \code{tidyplot()} function.}

\item{width}{bla}

\item{reverse}{bla}

\item{...}{bla}
}
\description{
Add pie or donut chart
}
