% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_sum_bar}
\alias{add_sum_bar}
\alias{add_sum_dash}
\alias{add_sum_dot}
\alias{add_sum_value}
\title{Add sum}
\usage{
add_sum_bar(gg, dodge_width = NULL, alpha = 1, width = 0.6, ...)

add_sum_dash(gg, dodge_width = NULL, width = 0.6, linewidth = 0.25, ...)

add_sum_dot(gg, dodge_width = NULL, size = 2, ...)

add_sum_value(
  gg,
  accuracy = 0.1,
  dodge_width = NULL,
  fontsize = 7,
  vjust = -0.5,
  padding_top = 0.15,
  ...
)
}
\arguments{
\item{gg}{bla}

\item{dodge_width}{bla}

\item{alpha}{description}

\item{width}{bla}

\item{...}{description}

\item{linewidth}{bla}

\item{size}{description}

\item{accuracy}{description}

\item{fontsize}{description}

\item{vjust}{description}

\item{padding_top}{description}
}
\description{
Add sum
}
