% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_text_labels}
\alias{add_text_labels}
\title{Add text labels}
\usage{
add_text_labels(
  plot,
  var,
  data = all_rows(),
  fontsize = 7,
  segment.size = 0.2,
  box.padding = 0.2,
  ...
)
}
\arguments{
\item{plot}{A tidyplot generated with the \code{tidyplot()} function.}

\item{var}{bla}

\item{data}{A function to subset the data to be plotted. See \code{filter_rows()} and friends.}

\item{fontsize}{bla}

\item{segment.size}{bla}

\item{box.padding}{bla}

\item{...}{bla}
}
\description{
Add text labels
}
