% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_title}
\alias{add_title}
\alias{add_caption}
\title{Add plot title or caption}
\usage{
add_title(plot, title = ggplot2::waiver())

add_caption(plot, caption = ggplot2::waiver())
}
\arguments{
\item{plot}{A tidyplot generated with the \code{tidyplot()} function.}

\item{title}{bla}

\item{caption}{bla}
}
\description{
Add plot title or caption
}
