% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust_description}
\alias{adjust_description}
\title{Adjust description}
\usage{
adjust_description(
  plot,
  title = ggplot2::waiver(),
  x_axis_title = ggplot2::waiver(),
  y_axis_title = ggplot2::waiver(),
  legend_title = ggplot2::waiver(),
  caption = ggplot2::waiver(),
  ...
)
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}

\item{title}{Plot title.}

\item{x_axis_title}{X axis title.}

\item{y_axis_title}{Y axis title.}

\item{legend_title}{Legend title.}

\item{caption}{Plot caption text.}

\item{...}{Arguments passed on to \code{ggplot2::labs()}.}
}
\value{
A \code{tidyplot} object
}
\description{
Adjust description
}
\details{
\itemize{
\item \code{adjust_description()} supports \href{https://www.rdocumentation.org/packages/grDevices/versions/3.6.2/topics/plotmath}{plotmath expressions} to include special characters.
See examples and \href{https://jbengler.github.io/tidyplots/articles/Advanced-plotting.html#special-characters}{Advanced plotting}.
}
}
\examples{
study \%>\%
  tidyplot(x = treatment, y = score, color = treatment) \%>\%
  add_data_points() \%>\%
  add_mean_bar(alpha = 0.3) \%>\%
  add_sem_bar() \%>\%
  adjust_description(
    title = "This is my fantastic plot title",
    x_axis_title = "Treatment group",
    y_axis_title = "Disease score",
    legend_title = "Legend title",
    caption = "Here goes the caption")
# Plotmath expressions
study \%>\%
  tidyplot(x = treatment, y = score, color = treatment) \%>\%
  add_data_points() \%>\%
  add_mean_bar(alpha = 0.3) \%>\%
  add_sem_bar() \%>\%
  adjust_description(
    title = "$H[2]*O$",
    x_axis_title = "$H[2]*O$",
    y_axis_title = "$H[2]*O$",
    legend_title = "$H[2]*O$",
    caption = "$H[2]*O$")

}
