% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust_description}
\alias{adjust_description}
\title{Adjust description}
\usage{
adjust_description(
  plot,
  title = ggplot2::waiver(),
  x_axis_title = ggplot2::waiver(),
  y_axis_title = ggplot2::waiver(),
  legend_title = ggplot2::waiver(),
  caption = ggplot2::waiver(),
  ...
)
}
\arguments{
\item{plot}{bla}

\item{title}{bla}

\item{x_axis_title}{bla}

\item{y_axis_title}{bla}

\item{legend_title}{bla}

\item{caption}{bla}

\item{...}{bla}
}
\description{
Adjust description
}
