% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust_font}
\alias{adjust_font}
\title{Adjust font}
\usage{
adjust_font(gg, fontsize = 7, family = NULL, face = NULL, color = "black")
}
\arguments{
\item{gg}{bla}

\item{fontsize}{bla}

\item{family}{Font family}

\item{face}{Font face ("plain", "italic", "bold", "bold.italic")}

\item{color}{bla}
}
\description{
Adjust font
}
