% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust_legend}
\alias{adjust_legend}
\title{Adjust legend}
\usage{
adjust_legend(plot, title = ggplot2::waiver(), position = "right")
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}

\item{title}{bla}

\item{position}{The position of legends. Can be one of \code{"none"}, \code{"left"}, \code{"right"},
\code{"bottom"}, \code{"top"}, or a two-element numeric vector.}
}
\value{
A \code{tidyplot} object
}
\description{
Adjust legend
}
