% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust_padding}
\alias{adjust_padding}
\title{Adjust padding}
\usage{
adjust_padding(
  plot,
  top = NA,
  right = NA,
  bottom = NA,
  left = NA,
  force_continuous = FALSE
)
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}

\item{top}{bla}

\item{right}{bla}

\item{bottom}{bla}

\item{left}{bla}

\item{force_continuous}{bla}
}
\value{
A \code{tidyplot} object
}
\description{
Adjust padding
}
