% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust_plot_area_padding}
\alias{adjust_plot_area_padding}
\title{Adjust plot area padding}
\usage{
adjust_plot_area_padding(
  plot,
  top = NA,
  right = NA,
  bottom = NA,
  left = NA,
  all = NA,
  force_continuous = FALSE,
  ...
)
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}

\item{top}{Extra space between the data points and the top. Defaults to \code{NA}, which does not change the padding.}

\item{right}{Extra space between the data points and the right. Defaults to \code{NA}, which does not change the padding.}

\item{bottom}{Extra space between the data points and the bottom. Defaults to \code{NA}, which does not change the padding.}

\item{left}{Extra space between the data points and the left. Defaults to \code{NA}, which does not change the padding.}

\item{all}{Extra space around the data pointst. Overwrites \code{top}, \code{right}, \code{bottom}, \code{left} if set. Defaults to \code{NA}, which does not change the padding.}

\item{force_continuous}{Whether to force the axis to be continuous. Defaults to \code{FALSE}.}

\item{...}{Arguments passed on to the \code{geom} function.}
}
\value{
A \code{tidyplot} object
}
\description{
Adjust plot area padding
}
