% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust_plot_size}
\alias{adjust_plot_size}
\title{Adjust plot size}
\usage{
adjust_plot_size(plot, width = 50, height = 50, unit = "mm")
}
\arguments{
\item{plot}{bla}

\item{width}{bla}

\item{height}{bla}

\item{unit}{bla}
}
\description{
Adjust plot size
}
