% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust_rotate_plot}
\alias{adjust_rotate_plot}
\title{Rotate plot by 90 degrees}
\usage{
adjust_rotate_plot(plot, ...)
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}

\item{...}{Arguments passed on to the \code{ggplot2::coord_flip()}.}
}
\value{
A \code{tidyplot} object
}
\description{
Rotate plot by 90 degrees
}
