% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust_x_axis}
\alias{adjust_x_axis}
\alias{adjust_y_axis}
\title{Adjust axes}
\usage{
adjust_x_axis(
  gg,
  title = ggplot2::waiver(),
  breaks = ggplot2::waiver(),
  labels = ggplot2::waiver(),
  limits = NULL,
  padding_left = NULL,
  padding_right = NULL,
  rotate_labels = FALSE,
  transformation = "identity",
  cut_short_scale = FALSE,
  position = "bottom",
  ...
)

adjust_y_axis(
  gg,
  title = ggplot2::waiver(),
  breaks = ggplot2::waiver(),
  labels = ggplot2::waiver(),
  limits = NULL,
  padding_bottom = NULL,
  padding_top = NULL,
  transformation = "identity",
  position = "left",
  cut_short_scale = FALSE,
  force_y_continuous = FALSE,
  ...
)
}
\arguments{
\item{gg}{bla}

\item{title}{bla}

\item{breaks}{bla}

\item{labels}{bla}

\item{limits}{bla}

\item{padding_left}{bla}

\item{padding_right}{bla}

\item{rotate_labels}{bla}

\item{transformation}{bla}

\item{cut_short_scale}{bla}

\item{position}{bla}

\item{...}{bla}

\item{padding_bottom}{bla}

\item{padding_top}{bla}

\item{force_y_continuous}{bla}
}
\description{
Adjust axes
}
