% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycolor.R
\docType{data}
\name{colors_continuous_viridis}
\alias{colors_continuous_viridis}
\alias{colors_continuous_magma}
\alias{colors_continuous_inferno}
\alias{colors_continuous_plasma}
\alias{colors_continuous_cividis}
\alias{colors_continuous_rocket}
\alias{colors_continuous_mako}
\alias{colors_continuous_turbo}
\alias{colors_continuous_bluepinkyellow}
\title{Continuous color schemes}
\format{
An object of class \code{tidycolor} (inherits from \code{character}) of length 265.

An object of class \code{tidycolor} (inherits from \code{character}) of length 265.

An object of class \code{tidycolor} (inherits from \code{character}) of length 265.

An object of class \code{tidycolor} (inherits from \code{character}) of length 265.

An object of class \code{tidycolor} (inherits from \code{character}) of length 265.

An object of class \code{tidycolor} (inherits from \code{character}) of length 265.

An object of class \code{tidycolor} (inherits from \code{character}) of length 265.

An object of class \code{tidycolor} (inherits from \code{character}) of length 265.

An object of class \code{tidycolor} (inherits from \code{character}) of length 11.
}
\usage{
colors_continuous_viridis

colors_continuous_magma

colors_continuous_inferno

colors_continuous_plasma

colors_continuous_cividis

colors_continuous_rocket

colors_continuous_mako

colors_continuous_turbo

colors_continuous_bluepinkyellow
}
\description{
For more information about the use of color schemes in tidyplots, check out this article:
\href{https://jbengler.github.io/tidyplots/articles/Color-schemes.html}{Color schemes}
}
\details{
Color schemes can be conveniently previewed by using the print method of the
\code{tidycolor} class. This will send a html preview to the RStudio Viewer pane.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_continuous_viridis
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_continuous_viridis</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 265 colors, downsampled to 42 colors.</small><br><span style='background-color:#440154FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#460A5DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#471264FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#481B6DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#482374FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#472C7AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#46337FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#443A83FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#424186FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3F4889FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3C508BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#39568CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#365D8DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#33638DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#306A8EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#2D708EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#2B758EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#297B8EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#26818EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#24878EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#228D8DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#20928CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#1F988BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#1F9F88FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#20A486FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#24AA83FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#29AF7FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#31B57BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3BBB75FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#45C06FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#53C569FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#5EC962FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#6ECE58FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#7BD250FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#8AD647FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#9CD93CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#AADC32FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#BDDF26FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#CCE11EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#DEE318FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#EDE51BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FDE725FF; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#440154FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#440154FF"</span>,<span style='background-color:#460A5DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#460A5DFF"</span>,<span style='background-color:#471264FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#471264FF"</span>,<span style='background-color:#481B6DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#481B6DFF"</span>,<span style='background-color:#482374FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#482374FF"</span>,<span style='background-color:#472C7AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#472C7AFF"</span>,<span style='background-color:#46337FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#46337FFF"</span>,<span style='background-color:#443A83FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#443A83FF"</span>,<span style='background-color:#424186FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#424186FF"</span>,<span style='background-color:#3F4889FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3F4889FF"</span>,<span style='background-color:#3C508BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3C508BFF"</span>,<span style='background-color:#39568CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#39568CFF"</span>,<span style='background-color:#365D8DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#365D8DFF"</span>,<span style='background-color:#33638DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#33638DFF"</span>,<span style='background-color:#306A8EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#306A8EFF"</span>,<span style='background-color:#2D708EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#2D708EFF"</span>,<span style='background-color:#2B758EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#2B758EFF"</span>,<span style='background-color:#297B8EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#297B8EFF"</span>,<span style='background-color:#26818EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#26818EFF"</span>,<span style='background-color:#24878EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#24878EFF"</span>,<span style='background-color:#228D8DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#228D8DFF"</span>,<span style='background-color:#20928CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#20928CFF"</span>,<span style='background-color:#1F988BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#1F988BFF"</span>,<span style='background-color:#1F9F88FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#1F9F88FF"</span>,<span style='background-color:#20A486FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#20A486FF"</span>,<span style='background-color:#24AA83FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#24AA83FF"</span>,<span style='background-color:#29AF7FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#29AF7FFF"</span>,<span style='background-color:#31B57BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#31B57BFF"</span>,<span style='background-color:#3BBB75FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3BBB75FF"</span>,<span style='background-color:#45C06FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#45C06FFF"</span>,<span style='background-color:#53C569FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#53C569FF"</span>,<span style='background-color:#5EC962FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#5EC962FF"</span>,<span style='background-color:#6ECE58FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#6ECE58FF"</span>,<span style='background-color:#7BD250FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#7BD250FF"</span>,<span style='background-color:#8AD647FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#8AD647FF"</span>,<span style='background-color:#9CD93CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#9CD93CFF"</span>,<span style='background-color:#AADC32FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#AADC32FF"</span>,<span style='background-color:#BDDF26FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#BDDF26FF"</span>,<span style='background-color:#CCE11EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#CCE11EFF"</span>,<span style='background-color:#DEE318FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#DEE318FF"</span>,<span style='background-color:#EDE51BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#EDE51BFF"</span>,<span style='background-color:#FDE725FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FDE725FF"</span>)</span></div><!--/html_preserve-->
}}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_continuous_magma
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_continuous_magma</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 265 colors, downsampled to 42 colors.</small><br><span style='background-color:#000004FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#02020DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#060519FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#0C0927FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#130D34FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#1C1044FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#231252FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#2E1162FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#38106CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#420F75FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4E117BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#57157EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#611880FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#6A1C81FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#752181FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#7D2482FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#882781FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#922B81FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#9B2E7FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#A6317DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#AF357BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#BB3978FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#C43C75FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#CF4070FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D8456CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#DF4B68FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E85362FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#EE5B5EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F4675CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F7715CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FA7E5EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FC8961FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FD9668FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FEA16EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FEAB75FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FEB87EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FEC287FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FECF92FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FDD99BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FDE6A8FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FCF0B2FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FCFDBFFF; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#000004FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#000004FF"</span>,<span style='background-color:#02020DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#02020DFF"</span>,<span style='background-color:#060519FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#060519FF"</span>,<span style='background-color:#0C0927FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#0C0927FF"</span>,<span style='background-color:#130D34FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#130D34FF"</span>,<span style='background-color:#1C1044FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#1C1044FF"</span>,<span style='background-color:#231252FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#231252FF"</span>,<span style='background-color:#2E1162FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#2E1162FF"</span>,<span style='background-color:#38106CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#38106CFF"</span>,<span style='background-color:#420F75FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#420F75FF"</span>,<span style='background-color:#4E117BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#4E117BFF"</span>,<span style='background-color:#57157EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#57157EFF"</span>,<span style='background-color:#611880FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#611880FF"</span>,<span style='background-color:#6A1C81FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#6A1C81FF"</span>,<span style='background-color:#752181FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#752181FF"</span>,<span style='background-color:#7D2482FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#7D2482FF"</span>,<span style='background-color:#882781FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#882781FF"</span>,<span style='background-color:#922B81FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#922B81FF"</span>,<span style='background-color:#9B2E7FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#9B2E7FFF"</span>,<span style='background-color:#A6317DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#A6317DFF"</span>,<span style='background-color:#AF357BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#AF357BFF"</span>,<span style='background-color:#BB3978FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#BB3978FF"</span>,<span style='background-color:#C43C75FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#C43C75FF"</span>,<span style='background-color:#CF4070FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#CF4070FF"</span>,<span style='background-color:#D8456CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#D8456CFF"</span>,<span style='background-color:#DF4B68FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#DF4B68FF"</span>,<span style='background-color:#E85362FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#E85362FF"</span>,<span style='background-color:#EE5B5EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#EE5B5EFF"</span>,<span style='background-color:#F4675CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F4675CFF"</span>,<span style='background-color:#F7715CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F7715CFF"</span>,<span style='background-color:#FA7E5EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FA7E5EFF"</span>,<span style='background-color:#FC8961FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FC8961FF"</span>,<span style='background-color:#FD9668FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FD9668FF"</span>,<span style='background-color:#FEA16EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FEA16EFF"</span>,<span style='background-color:#FEAB75FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FEAB75FF"</span>,<span style='background-color:#FEB87EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FEB87EFF"</span>,<span style='background-color:#FEC287FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FEC287FF"</span>,<span style='background-color:#FECF92FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FECF92FF"</span>,<span style='background-color:#FDD99BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FDD99BFF"</span>,<span style='background-color:#FDE6A8FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FDE6A8FF"</span>,<span style='background-color:#FCF0B2FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FCF0B2FF"</span>,<span style='background-color:#FCFDBFFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FCFDBFFF"</span>)</span></div><!--/html_preserve-->
}}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_continuous_inferno
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_continuous_inferno</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 265 colors, downsampled to 42 colors.</small><br><span style='background-color:#000004FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#02020EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#07051AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#0D082AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#150B37FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#1F0C48FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#290B54FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#350A60FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3E0966FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#490B6AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#540F6DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#5D126EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#68166EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#71196EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#7C1D6DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#85216BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#8F2568FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#992766FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#A22B62FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#AD305DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#B53458FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#C03952FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#C73E4CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D04545FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D84C3EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#DF5237FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E55C30FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#EB6429FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F06F20FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F47918FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F7840FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F98E09FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FB9B06FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FCA60CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FCB115FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FBBF24FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F9C932FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F6D746FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F3E259FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F1EE73FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F3F68BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FCFFA4FF; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#000004FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#000004FF"</span>,<span style='background-color:#02020EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#02020EFF"</span>,<span style='background-color:#07051AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#07051AFF"</span>,<span style='background-color:#0D082AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#0D082AFF"</span>,<span style='background-color:#150B37FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#150B37FF"</span>,<span style='background-color:#1F0C48FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#1F0C48FF"</span>,<span style='background-color:#290B54FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#290B54FF"</span>,<span style='background-color:#350A60FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#350A60FF"</span>,<span style='background-color:#3E0966FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3E0966FF"</span>,<span style='background-color:#490B6AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#490B6AFF"</span>,<span style='background-color:#540F6DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#540F6DFF"</span>,<span style='background-color:#5D126EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#5D126EFF"</span>,<span style='background-color:#68166EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#68166EFF"</span>,<span style='background-color:#71196EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#71196EFF"</span>,<span style='background-color:#7C1D6DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#7C1D6DFF"</span>,<span style='background-color:#85216BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#85216BFF"</span>,<span style='background-color:#8F2568FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#8F2568FF"</span>,<span style='background-color:#992766FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#992766FF"</span>,<span style='background-color:#A22B62FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#A22B62FF"</span>,<span style='background-color:#AD305DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#AD305DFF"</span>,<span style='background-color:#B53458FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#B53458FF"</span>,<span style='background-color:#C03952FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#C03952FF"</span>,<span style='background-color:#C73E4CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#C73E4CFF"</span>,<span style='background-color:#D04545FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#D04545FF"</span>,<span style='background-color:#D84C3EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#D84C3EFF"</span>,<span style='background-color:#DF5237FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#DF5237FF"</span>,<span style='background-color:#E55C30FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#E55C30FF"</span>,<span style='background-color:#EB6429FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#EB6429FF"</span>,<span style='background-color:#F06F20FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F06F20FF"</span>,<span style='background-color:#F47918FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F47918FF"</span>,<span style='background-color:#F7840FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F7840FFF"</span>,<span style='background-color:#F98E09FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F98E09FF"</span>,<span style='background-color:#FB9B06FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FB9B06FF"</span>,<span style='background-color:#FCA60CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FCA60CFF"</span>,<span style='background-color:#FCB115FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FCB115FF"</span>,<span style='background-color:#FBBF24FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FBBF24FF"</span>,<span style='background-color:#F9C932FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F9C932FF"</span>,<span style='background-color:#F6D746FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F6D746FF"</span>,<span style='background-color:#F3E259FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F3E259FF"</span>,<span style='background-color:#F1EE73FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F1EE73FF"</span>,<span style='background-color:#F3F68BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F3F68BFF"</span>,<span style='background-color:#FCFFA4FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FCFFA4FF"</span>)</span></div><!--/html_preserve-->
}}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_continuous_plasma
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_continuous_plasma</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 265 colors, downsampled to 42 colors.</small><br><span style='background-color:#0D0887FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#1C068EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#290593FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#360498FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3F049CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4B03A1FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#5502A4FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#5F01A6FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#6700A8FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#7100A8FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#7B02A8FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#8405A7FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#8E0BA5FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#9511A1FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#9E199DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#A62098FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#AD2792FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#B42E8DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#BB3488FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#C23C81FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#C8437BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#CD4A76FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D35171FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D9586AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#DE5F65FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E26560FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E76E5BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#EB7556FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#EF7E50FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F3854BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F68E44FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F89540FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FB9F3AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FCA835FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FDB030FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FEBA2CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FDC328FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FCCE25FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FAD824FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F7E325FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F4ED27FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F0F921FF; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#0D0887FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#0D0887FF"</span>,<span style='background-color:#1C068EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#1C068EFF"</span>,<span style='background-color:#290593FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#290593FF"</span>,<span style='background-color:#360498FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#360498FF"</span>,<span style='background-color:#3F049CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3F049CFF"</span>,<span style='background-color:#4B03A1FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#4B03A1FF"</span>,<span style='background-color:#5502A4FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#5502A4FF"</span>,<span style='background-color:#5F01A6FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#5F01A6FF"</span>,<span style='background-color:#6700A8FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#6700A8FF"</span>,<span style='background-color:#7100A8FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#7100A8FF"</span>,<span style='background-color:#7B02A8FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#7B02A8FF"</span>,<span style='background-color:#8405A7FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#8405A7FF"</span>,<span style='background-color:#8E0BA5FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#8E0BA5FF"</span>,<span style='background-color:#9511A1FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#9511A1FF"</span>,<span style='background-color:#9E199DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#9E199DFF"</span>,<span style='background-color:#A62098FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#A62098FF"</span>,<span style='background-color:#AD2792FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#AD2792FF"</span>,<span style='background-color:#B42E8DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#B42E8DFF"</span>,<span style='background-color:#BB3488FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#BB3488FF"</span>,<span style='background-color:#C23C81FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#C23C81FF"</span>,<span style='background-color:#C8437BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#C8437BFF"</span>,<span style='background-color:#CD4A76FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#CD4A76FF"</span>,<span style='background-color:#D35171FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#D35171FF"</span>,<span style='background-color:#D9586AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#D9586AFF"</span>,<span style='background-color:#DE5F65FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#DE5F65FF"</span>,<span style='background-color:#E26560FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#E26560FF"</span>,<span style='background-color:#E76E5BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#E76E5BFF"</span>,<span style='background-color:#EB7556FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#EB7556FF"</span>,<span style='background-color:#EF7E50FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#EF7E50FF"</span>,<span style='background-color:#F3854BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F3854BFF"</span>,<span style='background-color:#F68E44FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F68E44FF"</span>,<span style='background-color:#F89540FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F89540FF"</span>,<span style='background-color:#FB9F3AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FB9F3AFF"</span>,<span style='background-color:#FCA835FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FCA835FF"</span>,<span style='background-color:#FDB030FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FDB030FF"</span>,<span style='background-color:#FEBA2CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FEBA2CFF"</span>,<span style='background-color:#FDC328FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FDC328FF"</span>,<span style='background-color:#FCCE25FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FCCE25FF"</span>,<span style='background-color:#FAD824FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FAD824FF"</span>,<span style='background-color:#F7E325FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F7E325FF"</span>,<span style='background-color:#F4ED27FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F4ED27FF"</span>,<span style='background-color:#F0F921FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F0F921FF"</span>)</span></div><!--/html_preserve-->
}}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_continuous_cividis
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_continuous_cividis</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 265 colors, downsampled to 42 colors.</small><br><span style='background-color:#00204DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#002557FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#002961FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#002D6DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#00306FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#01366EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#173A6DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#253E6CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#2E436CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#36476BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3F4C6BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#45506BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4C546CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#52596CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#595E6DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#5E626EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#64666FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#696B71FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#6E6F73FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#747475FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#797977FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#7F7D78FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#848279FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#8B8779FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#918C78FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#979178FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#9D9677FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#A39A76FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#ABA074FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#B1A573FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#B7AA71FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#BEAF6FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#C5B56CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#CCBB69FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D2C066FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D9C562FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E0CB5EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E8D259FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#EED753FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F7DD4DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FDE346FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FFEA46FF; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#00204DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#00204DFF"</span>,<span style='background-color:#002557FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#002557FF"</span>,<span style='background-color:#002961FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#002961FF"</span>,<span style='background-color:#002D6DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#002D6DFF"</span>,<span style='background-color:#00306FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#00306FFF"</span>,<span style='background-color:#01366EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#01366EFF"</span>,<span style='background-color:#173A6DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#173A6DFF"</span>,<span style='background-color:#253E6CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#253E6CFF"</span>,<span style='background-color:#2E436CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#2E436CFF"</span>,<span style='background-color:#36476BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#36476BFF"</span>,<span style='background-color:#3F4C6BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3F4C6BFF"</span>,<span style='background-color:#45506BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#45506BFF"</span>,<span style='background-color:#4C546CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#4C546CFF"</span>,<span style='background-color:#52596CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#52596CFF"</span>,<span style='background-color:#595E6DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#595E6DFF"</span>,<span style='background-color:#5E626EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#5E626EFF"</span>,<span style='background-color:#64666FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#64666FFF"</span>,<span style='background-color:#696B71FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#696B71FF"</span>,<span style='background-color:#6E6F73FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#6E6F73FF"</span>,<span style='background-color:#747475FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#747475FF"</span>,<span style='background-color:#797977FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#797977FF"</span>,<span style='background-color:#7F7D78FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#7F7D78FF"</span>,<span style='background-color:#848279FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#848279FF"</span>,<span style='background-color:#8B8779FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#8B8779FF"</span>,<span style='background-color:#918C78FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#918C78FF"</span>,<span style='background-color:#979178FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#979178FF"</span>,<span style='background-color:#9D9677FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#9D9677FF"</span>,<span style='background-color:#A39A76FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#A39A76FF"</span>,<span style='background-color:#ABA074FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#ABA074FF"</span>,<span style='background-color:#B1A573FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#B1A573FF"</span>,<span style='background-color:#B7AA71FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#B7AA71FF"</span>,<span style='background-color:#BEAF6FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#BEAF6FFF"</span>,<span style='background-color:#C5B56CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#C5B56CFF"</span>,<span style='background-color:#CCBB69FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#CCBB69FF"</span>,<span style='background-color:#D2C066FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#D2C066FF"</span>,<span style='background-color:#D9C562FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#D9C562FF"</span>,<span style='background-color:#E0CB5EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#E0CB5EFF"</span>,<span style='background-color:#E8D259FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#E8D259FF"</span>,<span style='background-color:#EED753FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#EED753FF"</span>,<span style='background-color:#F7DD4DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F7DD4DFF"</span>,<span style='background-color:#FDE346FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FDE346FF"</span>,<span style='background-color:#FFEA46FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FFEA46FF"</span>)</span></div><!--/html_preserve-->
}}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_continuous_rocket
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_continuous_rocket</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 265 colors, downsampled to 42 colors.</small><br><span style='background-color:#03051AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#0A091FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#120D25FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#1C112BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#241432FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#2E1739FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#36193FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#411B44FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#491D49FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#531E4DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#5E1F52FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#671F55FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#721F57FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#7B1F59FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#871E5BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#921C5BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#9D1B5BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#A7195AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#B01759FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#BC1656FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#C51852FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#CE1D4EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D62449FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#DE2E44FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E43841FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E8413EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#ED4F3EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#EF5A41FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F26747FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F3724EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F47E57FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F58860FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F5946BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F69D75FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F6A77FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F6B28CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F6BB97FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F7C6A6FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F7CEB2FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F8D8C1FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F9E0CEFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FAEBDDFF; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#03051AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#03051AFF"</span>,<span style='background-color:#0A091FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#0A091FFF"</span>,<span style='background-color:#120D25FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#120D25FF"</span>,<span style='background-color:#1C112BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#1C112BFF"</span>,<span style='background-color:#241432FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#241432FF"</span>,<span style='background-color:#2E1739FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#2E1739FF"</span>,<span style='background-color:#36193FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#36193FFF"</span>,<span style='background-color:#411B44FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#411B44FF"</span>,<span style='background-color:#491D49FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#491D49FF"</span>,<span style='background-color:#531E4DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#531E4DFF"</span>,<span style='background-color:#5E1F52FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#5E1F52FF"</span>,<span style='background-color:#671F55FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#671F55FF"</span>,<span style='background-color:#721F57FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#721F57FF"</span>,<span style='background-color:#7B1F59FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#7B1F59FF"</span>,<span style='background-color:#871E5BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#871E5BFF"</span>,<span style='background-color:#921C5BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#921C5BFF"</span>,<span style='background-color:#9D1B5BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#9D1B5BFF"</span>,<span style='background-color:#A7195AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#A7195AFF"</span>,<span style='background-color:#B01759FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#B01759FF"</span>,<span style='background-color:#BC1656FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#BC1656FF"</span>,<span style='background-color:#C51852FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#C51852FF"</span>,<span style='background-color:#CE1D4EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#CE1D4EFF"</span>,<span style='background-color:#D62449FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#D62449FF"</span>,<span style='background-color:#DE2E44FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#DE2E44FF"</span>,<span style='background-color:#E43841FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#E43841FF"</span>,<span style='background-color:#E8413EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#E8413EFF"</span>,<span style='background-color:#ED4F3EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#ED4F3EFF"</span>,<span style='background-color:#EF5A41FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#EF5A41FF"</span>,<span style='background-color:#F26747FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F26747FF"</span>,<span style='background-color:#F3724EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F3724EFF"</span>,<span style='background-color:#F47E57FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F47E57FF"</span>,<span style='background-color:#F58860FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F58860FF"</span>,<span style='background-color:#F5946BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F5946BFF"</span>,<span style='background-color:#F69D75FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F69D75FF"</span>,<span style='background-color:#F6A77FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F6A77FFF"</span>,<span style='background-color:#F6B28CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F6B28CFF"</span>,<span style='background-color:#F6BB97FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F6BB97FF"</span>,<span style='background-color:#F7C6A6FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F7C6A6FF"</span>,<span style='background-color:#F7CEB2FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F7CEB2FF"</span>,<span style='background-color:#F8D8C1FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F8D8C1FF"</span>,<span style='background-color:#F9E0CEFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F9E0CEFF"</span>,<span style='background-color:#FAEBDDFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FAEBDDFF"</span>)</span></div><!--/html_preserve-->
}}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_continuous_mako
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_continuous_mako</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 265 colors, downsampled to 42 colors.</small><br><span style='background-color:#0B0405FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#12080DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#180D16FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#1E111FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#241628FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#2A1B33FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#2F1F3DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#342547FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#372852FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3B2D5BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3E3367FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#403871FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#413E7EFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#414387FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3F4B90FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3D5296FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3A5A9AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#38619DFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#37689FFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3670A0FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3576A2FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#357DA3FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3484A5FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#348CA7FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3492A8FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3499AAFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#35A0ABFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#37A6ACFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3AAEADFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3FB5ADFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#44BDADFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4BC2ADFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#55CAADFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#61CFACFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#72D4ADFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#86D9B1FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#96DDB5FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#A9E1BDFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#B6E6C5FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#C5EAD0FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D0EFDAFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#DEF5E5FF; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#0B0405FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#0B0405FF"</span>,<span style='background-color:#12080DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#12080DFF"</span>,<span style='background-color:#180D16FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#180D16FF"</span>,<span style='background-color:#1E111FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#1E111FFF"</span>,<span style='background-color:#241628FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#241628FF"</span>,<span style='background-color:#2A1B33FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#2A1B33FF"</span>,<span style='background-color:#2F1F3DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#2F1F3DFF"</span>,<span style='background-color:#342547FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#342547FF"</span>,<span style='background-color:#372852FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#372852FF"</span>,<span style='background-color:#3B2D5BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3B2D5BFF"</span>,<span style='background-color:#3E3367FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3E3367FF"</span>,<span style='background-color:#403871FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#403871FF"</span>,<span style='background-color:#413E7EFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#413E7EFF"</span>,<span style='background-color:#414387FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#414387FF"</span>,<span style='background-color:#3F4B90FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3F4B90FF"</span>,<span style='background-color:#3D5296FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3D5296FF"</span>,<span style='background-color:#3A5A9AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3A5A9AFF"</span>,<span style='background-color:#38619DFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#38619DFF"</span>,<span style='background-color:#37689FFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#37689FFF"</span>,<span style='background-color:#3670A0FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3670A0FF"</span>,<span style='background-color:#3576A2FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3576A2FF"</span>,<span style='background-color:#357DA3FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#357DA3FF"</span>,<span style='background-color:#3484A5FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3484A5FF"</span>,<span style='background-color:#348CA7FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#348CA7FF"</span>,<span style='background-color:#3492A8FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3492A8FF"</span>,<span style='background-color:#3499AAFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3499AAFF"</span>,<span style='background-color:#35A0ABFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#35A0ABFF"</span>,<span style='background-color:#37A6ACFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#37A6ACFF"</span>,<span style='background-color:#3AAEADFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3AAEADFF"</span>,<span style='background-color:#3FB5ADFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3FB5ADFF"</span>,<span style='background-color:#44BDADFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#44BDADFF"</span>,<span style='background-color:#4BC2ADFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#4BC2ADFF"</span>,<span style='background-color:#55CAADFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#55CAADFF"</span>,<span style='background-color:#61CFACFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#61CFACFF"</span>,<span style='background-color:#72D4ADFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#72D4ADFF"</span>,<span style='background-color:#86D9B1FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#86D9B1FF"</span>,<span style='background-color:#96DDB5FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#96DDB5FF"</span>,<span style='background-color:#A9E1BDFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#A9E1BDFF"</span>,<span style='background-color:#B6E6C5FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#B6E6C5FF"</span>,<span style='background-color:#C5EAD0FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#C5EAD0FF"</span>,<span style='background-color:#D0EFDAFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#D0EFDAFF"</span>,<span style='background-color:#DEF5E5FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#DEF5E5FF"</span>)</span></div><!--/html_preserve-->
}}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_continuous_turbo
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_continuous_turbo</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 265 colors, downsampled to 42 colors.</small><br><span style='background-color:#30123BFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#372365FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3D3489FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4147AEFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4456C8FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4669E0FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4777EFFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4688FBFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#4197FFFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#38A5FBFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#2CB7F0FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#22C4E3FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#1AD3D1FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#18DDC2FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#1DE7B2FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#29EFA2FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#3DF58CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#53FA79FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#69FD66FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#83FF51FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#98FE43FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#AAFB39FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#BAF635FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#CBED34FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D9E436FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E4DA38FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F0CC3AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F7C13AFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FCB136FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FEA230FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FE8F28FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FB7D21FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F56918FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#EF5911FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E74A0CFF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#DD3C08FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#D23105FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#C32503FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#B51C01FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#A31301FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#910B01FF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#7A0403FF; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#30123BFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#30123BFF"</span>,<span style='background-color:#372365FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#372365FF"</span>,<span style='background-color:#3D3489FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3D3489FF"</span>,<span style='background-color:#4147AEFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#4147AEFF"</span>,<span style='background-color:#4456C8FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#4456C8FF"</span>,<span style='background-color:#4669E0FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#4669E0FF"</span>,<span style='background-color:#4777EFFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#4777EFFF"</span>,<span style='background-color:#4688FBFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#4688FBFF"</span>,<span style='background-color:#4197FFFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#4197FFFF"</span>,<span style='background-color:#38A5FBFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#38A5FBFF"</span>,<span style='background-color:#2CB7F0FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#2CB7F0FF"</span>,<span style='background-color:#22C4E3FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#22C4E3FF"</span>,<span style='background-color:#1AD3D1FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#1AD3D1FF"</span>,<span style='background-color:#18DDC2FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#18DDC2FF"</span>,<span style='background-color:#1DE7B2FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#1DE7B2FF"</span>,<span style='background-color:#29EFA2FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#29EFA2FF"</span>,<span style='background-color:#3DF58CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#3DF58CFF"</span>,<span style='background-color:#53FA79FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#53FA79FF"</span>,<span style='background-color:#69FD66FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#69FD66FF"</span>,<span style='background-color:#83FF51FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#83FF51FF"</span>,<span style='background-color:#98FE43FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#98FE43FF"</span>,<span style='background-color:#AAFB39FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#AAFB39FF"</span>,<span style='background-color:#BAF635FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#BAF635FF"</span>,<span style='background-color:#CBED34FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#CBED34FF"</span>,<span style='background-color:#D9E436FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#D9E436FF"</span>,<span style='background-color:#E4DA38FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#E4DA38FF"</span>,<span style='background-color:#F0CC3AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F0CC3AFF"</span>,<span style='background-color:#F7C13AFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F7C13AFF"</span>,<span style='background-color:#FCB136FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FCB136FF"</span>,<span style='background-color:#FEA230FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FEA230FF"</span>,<span style='background-color:#FE8F28FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FE8F28FF"</span>,<span style='background-color:#FB7D21FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#FB7D21FF"</span>,<span style='background-color:#F56918FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#F56918FF"</span>,<span style='background-color:#EF5911FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#EF5911FF"</span>,<span style='background-color:#E74A0CFF; display: inline-block; padding: 2px 10px 2px 10px;'>"#E74A0CFF"</span>,<span style='background-color:#DD3C08FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#DD3C08FF"</span>,<span style='background-color:#D23105FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#D23105FF"</span>,<span style='background-color:#C32503FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#C32503FF"</span>,<span style='background-color:#B51C01FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#B51C01FF"</span>,<span style='background-color:#A31301FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#A31301FF"</span>,<span style='background-color:#910B01FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#910B01FF"</span>,<span style='background-color:#7A0403FF; display: inline-block; padding: 2px 10px 2px 10px;'>"#7A0403FF"</span>)</span></div><!--/html_preserve-->
}}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{colors_continuous_bluepinkyellow
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve--><div style="margin-top: 60px; margin-bottom: 60px;"><h4 style="font-family: 'Courier New', monospace;">colors_continuous_bluepinkyellow</h4><small style="font-family: 'Courier New', monospace;">A tidyplots color scheme with 11 colors.</small><br><span style='background-color:#00034D; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#000F9F; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#001CEF; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#241EF5; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#5823F6; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#A033E0; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#E85AB1; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F1907C; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#F4AF63; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FCE552; display: inline-block; width:30px; height:20px;'></span><span style='background-color:#FFFB6D; display: inline-block; width:30px; height:20px;'></span><br><span style="font-family: 'Courier New', monospace;">c(<br><span style='background-color:#00034D; display: inline-block; padding: 2px 10px 2px 10px;'>"#00034D"</span>,<span style='background-color:#000F9F; display: inline-block; padding: 2px 10px 2px 10px;'>"#000F9F"</span>,<span style='background-color:#001CEF; display: inline-block; padding: 2px 10px 2px 10px;'>"#001CEF"</span>,<span style='background-color:#241EF5; display: inline-block; padding: 2px 10px 2px 10px;'>"#241EF5"</span>,<span style='background-color:#5823F6; display: inline-block; padding: 2px 10px 2px 10px;'>"#5823F6"</span>,<span style='background-color:#A033E0; display: inline-block; padding: 2px 10px 2px 10px;'>"#A033E0"</span>,<span style='background-color:#E85AB1; display: inline-block; padding: 2px 10px 2px 10px;'>"#E85AB1"</span>,<span style='background-color:#F1907C; display: inline-block; padding: 2px 10px 2px 10px;'>"#F1907C"</span>,<span style='background-color:#F4AF63; display: inline-block; padding: 2px 10px 2px 10px;'>"#F4AF63"</span>,<span style='background-color:#FCE552; display: inline-block; padding: 2px 10px 2px 10px;'>"#FCE552"</span>,<span style='background-color:#FFFB6D; display: inline-block; padding: 2px 10px 2px 10px;'>"#FFFB6D"</span>)</span></div><!--/html_preserve-->
}}
}
\keyword{datasets}
