% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dinosaurs}
\alias{dinosaurs}
\title{Dinosaurs data}
\format{
A data frame.
}
\source{
ChatGPT-3.5, \emph{Caution: The accuracy of the data has not been verified.}
}
\usage{
dinosaurs
}
\description{
Dinosaurs data
}
\examples{
dplyr::glimpse(dinosaurs)
}
\keyword{datasets}
