% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distributions}
\alias{distributions}
\title{Distributions data}
\format{
A data frame.
}
\source{
tidyplots package
}
\usage{
distributions
}
\description{
Distributions data
}
\examples{
dplyr::glimpse(distributions)
}
\keyword{datasets}
