% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{energy}
\alias{energy}
\title{Energy data}
\format{
A data frame.
}
\source{
\href{https://www.energy-charts.info}{Energy-Charts}, Energy production data, Germany
}
\usage{
energy
}
\description{
Energy data
}
\examples{
dplyr::glimpse(energy)
}
\keyword{datasets}
