% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{energy_week}
\alias{energy_week}
\title{Energy week data}
\format{
A data frame.
}
\source{
\href{https://www.energy-charts.info}{Energy-Charts}, Energy production data, Germany
}
\usage{
energy_week
}
\description{
Energy week data
}
\examples{
dplyr::glimpse(energy_week)
}
\keyword{datasets}
