% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eu_countries}
\alias{eu_countries}
\title{EU countries data}
\format{
A data frame.
}
\source{
\code{tidyplots} package
}
\usage{
eu_countries
}
\description{
EU countries data
}
\examples{
dplyr::glimpse(eu_countries)
}
\keyword{datasets}
