% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_expression}
\alias{gene_expression}
\title{RNA-Seq expression data}
\format{
A data frame.
}
\source{
\href{https://www.nature.com/articles/s41593-019-0385-4}{Bassoon proteinopathy drives neurodegeneration in multiple sclerosis}, Nature Neuroscience

\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE104899}{GSE104899}, Gene Expression Omnibus
}
\usage{
gene_expression
}
\description{
RNA-Seq expression data
}
\examples{
dplyr::glimpse(gene_expression)
}
\keyword{datasets}
