% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{new_color_scheme}
\alias{new_color_scheme}
\title{New color scheme}
\usage{
new_color_scheme(x, name = "Untitled color scheme")
}
\arguments{
\item{x}{bla}

\item{name}{bla}
}
\description{
New color scheme
}
