% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycolor.R
\name{new_color_scheme}
\alias{new_color_scheme}
\title{New color scheme}
\usage{
new_color_scheme(x, name = "Untitled color scheme", reverse = FALSE)
}
\arguments{
\item{x}{Character vector of hex colors. For example \code{x = c("#FF00FF", "#00FFFF")}.}

\item{name}{Name of the custom color scheme.}

\item{reverse}{Whether the order should be reversed or not. Defaults to \code{FALSE}, meaning not reversed.}
}
\value{
A \code{tidyplot} object.
}
\description{
For more information about the use of color schemes in tidyplots, check out this article:
\href{https://jbengler.github.io/tidyplots/articles/Color-schemes.html}{Color schemes}
}
\examples{
new_color_scheme(c("#ECA669","#E06681","#8087E2","#E2D269"))

new_color_scheme(c("#ECA669","#E06681","#8087E2","#E2D269"),
  name = "my_custom_color_scheme")

}
