% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_legend}
\alias{remove_legend}
\alias{remove_legend_title}
\title{Remove legend or legend title}
\usage{
remove_legend(plot)

remove_legend_title(plot)
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}
}
\value{
A \code{tidyplot} object.
}
\description{
Remove legend or legend title
}
\examples{
# Before removing
study \%>\%
  tidyplot(x = treatment, y = score, color = treatment) \%>\%
  add_mean_bar()

# After removing
study \%>\%
  tidyplot(x = treatment, y = score, color = treatment) \%>\%
  add_mean_bar() \%>\%
  remove_legend_title()

study \%>\%
  tidyplot(x = treatment, y = score, color = treatment) \%>\%
  add_mean_bar() \%>\%
  remove_legend()

}
