% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_padding}
\alias{remove_padding}
\title{Remove padding}
\usage{
remove_padding(plot, force_continuous = FALSE)
}
\arguments{
\item{plot}{bla}

\item{force_continuous}{bla}
}
\description{
Remove padding
}
