% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_padding}
\alias{remove_padding}
\title{Remove plot area padding}
\usage{
remove_padding(plot, force_continuous = FALSE)
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}

\item{force_continuous}{Whether to force the axis to be continuous. Defaults to \code{FALSE}.}
}
\value{
A \code{tidyplot} object.
}
\description{
Remove plot area padding
}
\examples{
# Before removing
animals \%>\%
  tidyplot(x = weight, y = speed, color = family) \%>\%
  add_data_points()

# After removing
animals \%>\%
  tidyplot(x = weight, y = speed, color = family) \%>\%
  add_data_points() \%>\%
  remove_padding()

}
