% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_plot_area_padding}
\alias{remove_plot_area_padding}
\title{Remove plot area padding}
\usage{
remove_plot_area_padding(plot, force_continuous = FALSE)
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}

\item{force_continuous}{Whether to force the axis to be continuous. Defaults to \code{FALSE}.}
}
\value{
A \code{tidyplot} object
}
\description{
Remove plot area padding
}
\examples{
study \%>\%
  tidyplot(x = treatment, y = score, color = treatment) \%>\%
  add_mean_bar() \%>\%
  remove_plot_area_padding()
# Before removing plot area padding
animals \%>\%
  tidyplot(x = weight, y = speed, color = family) \%>\%
  add_data_points()
# After removing plot area padding
animals \%>\%
  tidyplot(x = weight, y = speed, color = family) \%>\%
  add_data_points() \%>\%
  remove_plot_area_padding()

}
