% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_x_axis}
\alias{remove_x_axis}
\alias{remove_x_axis_line}
\alias{remove_x_axis_ticks}
\alias{remove_x_axis_labels}
\alias{remove_x_axis_title}
\title{Remove x axis or parts of it}
\usage{
remove_x_axis(plot)

remove_x_axis_line(plot)

remove_x_axis_ticks(plot)

remove_x_axis_labels(plot)

remove_x_axis_title(plot)
}
\arguments{
\item{plot}{A \code{tidyplot} generated with the function \code{tidyplot()}.}
}
\value{
A \code{tidyplot} object
}
\description{
Remove x axis or parts of it
}
