% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_y_axis}
\alias{remove_y_axis}
\alias{remove_y_axis_line}
\alias{remove_y_axis_ticks}
\alias{remove_y_axis_labels}
\alias{remove_y_axis_title}
\title{Remove \code{y} axis or parts of it}
\usage{
remove_y_axis(gg)

remove_y_axis_line(gg)

remove_y_axis_ticks(gg)

remove_y_axis_labels(gg)

remove_y_axis_title(gg)
}
\arguments{
\item{gg}{bla}
}
\description{
Remove \code{y} axis or parts of it
}
