% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{rename_x_axis_labels}
\alias{rename_x_axis_labels}
\alias{rename_y_axis_labels}
\alias{rename_color_labels}
\title{Rename axis and color labels}
\usage{
rename_x_axis_labels(gg, new_names)

rename_y_axis_labels(gg, new_names)

rename_color_labels(gg, new_names)
}
\arguments{
\item{gg}{bla}

\item{new_names}{bla}
}
\description{
Rename axis and color labels
}
