% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{reorder_x_axis_labels}
\alias{reorder_x_axis_labels}
\alias{reorder_y_axis_labels}
\title{Reorder axis labels}
\usage{
reorder_x_axis_labels(gg, ...)

reorder_y_axis_labels(gg, ...)
}
\arguments{
\item{gg}{bla}

\item{...}{bla}
}
\description{
Reorder axis labels
}
