% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{reverse_x_axis_labels}
\alias{reverse_x_axis_labels}
\alias{reverse_y_axis_labels}
\alias{reverse_color_labels}
\title{Reverse axis or color labels}
\usage{
reverse_x_axis_labels(plot)

reverse_y_axis_labels(plot)

reverse_color_labels(plot)
}
\arguments{
\item{plot}{bla}
}
\description{
Reverse axis or color labels
}
