% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{sort_x_axis_labels}
\alias{sort_x_axis_labels}
\alias{sort_y_axis_labels}
\alias{sort_color_labels}
\title{Sort axis and color labels}
\usage{
sort_x_axis_labels(gg, ...)

sort_y_axis_labels(gg, ...)

sort_color_labels(gg, ...)
}
\arguments{
\item{gg}{bla}

\item{...}{bla}
}
\description{
Sort axis and color labels
}
