% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{sort_x_axis_labels}
\alias{sort_x_axis_labels}
\alias{sort_y_axis_labels}
\alias{sort_color_labels}
\title{Sort axis or color labels}
\usage{
sort_x_axis_labels(plot, ...)

sort_y_axis_labels(plot, ...)

sort_color_labels(plot, ...)
}
\arguments{
\item{plot}{bla}

\item{...}{bla}
}
\description{
Sort axis or color labels
}
