% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spendings}
\alias{spendings}
\title{Spending data}
\format{
A data frame.
}
\source{
tidyplots package
}
\usage{
spendings
}
\description{
Spending data
}
\examples{
dplyr::glimpse(spendings)
}
\keyword{datasets}
