% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_tidyplot}
\alias{theme_tidyplot}
\alias{theme_ggplot2}
\alias{theme_minimal_xy}
\alias{theme_minimal_y}
\alias{theme_minimal_x}
\title{Themes}
\usage{
theme_tidyplot(gg, fontsize = 7)

theme_ggplot2(gg, fontsize = 7)

theme_minimal_xy(gg, fontsize = 7)

theme_minimal_y(gg, fontsize = 7)

theme_minimal_x(gg, fontsize = 7)
}
\arguments{
\item{gg}{bla}

\item{fontsize}{bla}
}
\description{
Themes
}
