% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tidyplot}
\alias{tidyplot}
\title{Create a new tidyplot}
\usage{
tidyplot(data, ..., width = 50, height = 50, dodge_width = 0.8)
}
\arguments{
\item{data}{bla}

\item{...}{bla}

\item{width}{bla}

\item{height}{bla}

\item{dodge_width}{bla}
}
\description{
Create a new tidyplot
}
