% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyplots-package.R
\docType{package}
\name{tidyplots-package}
\alias{tidyplots}
\alias{tidyplots-package}
\title{tidyplots: Tidy Plots for Scientific Papers}
\description{
The goal of 'tidyplots' is to streamline the creation of publication-ready plots for scientific papers. It allows to gradually add, remove and adjust plot components using a consistent and intuitive syntax.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jbengler/tidyplots}
  \item \url{https://jbengler.github.io/tidyplots/}
  \item Report bugs at \url{https://github.com/jbengler/tidyplots/issues}
}

}
\author{
\strong{Maintainer}: Jan Broder Engler \email{broder.engler@gmail.com} (\href{https://orcid.org/0000-0002-3169-2076}{ORCID}) [copyright holder]

}
\keyword{internal}
