% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time_course}
\alias{time_course}
\title{Time course data}
\format{
A data frame.
}
\source{
tidyplots package
}
\usage{
time_course
}
\description{
Time course data
}
\examples{
dplyr::glimpse(time_course)
}
\keyword{datasets}
