% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{view_plot}
\alias{view_plot}
\title{View plot on screen}
\usage{
view_plot(plot, data = all_rows(), title = ggplot2::waiver(), ...)
}
\arguments{
\item{plot}{bla}

\item{data}{bla}

\item{title}{bla}

\item{...}{bla}
}
\description{
View plot on screen
}
